cleanup in models
