--- envinorma/models/arretes_ministeriels.py
@@ -1,2734 +1,0 @@
-ARRETES_MINISTERIELS = [
-    {
-        "aida_page": "35346",
-        "title": " Arrêté du 23/07/97 relatif aux stockages de chlore gazeux liquéfié sous pression lorsque la quantité totale susceptible d'être présente dans l'installation est supérieure ou égale à 18 tonnes ",
-        "nor": "ATEP9760313A",
-        "classements": [{"regime": "A", "rubrique": "1138"}, {"regime": "A", "rubrique": "4710"}],
-        "cid": "JORFTEXT000000568479",
-        "state": "VIGUEUR",
-        "date_of_signature": 869608800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4111",
-        "title": " Arrêté du 15/04/10 fixant les règles générales et prescriptions techniques applicables aux stations-service soumises à autorisation sous la rubrique n° 1435 de la nomenclature des installations classées pour la protection de l'environnement (Abrogé) ",
-        "nor": "DEVP1001966A",
-        "classements": [{"regime": "A", "rubrique": "1435"}],
-        "cid": "JORFTEXT000022097828",
-        "state": "ABROGE",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35815",
-        "title": " Arrêté du 03/10/10 relatif au stockage en réservoirs aériens manufacturés exploités au sein d'une installation classée soumise à autorisation au titre de l'une ou plusieurs des rubriques n° 1436, 4330, 4331, 4722, 4734, 4742, 4743, 4744, 4746, 4747 ou 4748, ou pour le pétrole brut au titre de l'une ou plusieurs des rubriques n° 4510 ou 4511 (à compter du 1er janvier 2021, l'intitulé devient « relatif au stockage en réservoirs aériens manufacturés de liquides inflammables, exploités au sein d'une installation classée pour la protection de l'environnement soumise à autorisation ») ",
-        "nor": "DEVP1025848A",
-        "classements": [
-            {"regime": "A", "rubrique": "1432"},
-            {"regime": "A", "rubrique": "1436"},
-            {"regime": "A", "rubrique": "4330"},
-            {"regime": "A", "rubrique": "4331"},
-            {"regime": "A", "rubrique": "4510"},
-            {"regime": "A", "rubrique": "4511"},
-            {"regime": "A", "rubrique": "4722"},
-            {"regime": "A", "rubrique": "4734"},
-            {"regime": "A", "rubrique": "4742"},
-            {"regime": "A", "rubrique": "4743"},
-            {"regime": "A", "rubrique": "4744"},
-            {"regime": "A", "rubrique": "4746"},
-            {"regime": "A", "rubrique": "4747"},
-            {"regime": "A", "rubrique": "4748"},
-        ],
-        "cid": "JORFTEXT000023081678",
-        "state": "VIGUEUR",
-        "date_of_signature": 1286056800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35819",
-        "title": " Arrêté du 16/07/12 relatif aux stockages en récipients mobiles exploités au sein d'une installation classée soumise à autorisation au titre de l'une ou plusieurs des rubriques n° 1436, 4330, 4331, 4722, 4734, 4742, 4743, 4744, 4746, 4747 ou 4748, ou pour le pétrole brut au titre de l'une ou plusieurs des rubriques n° 4510 ou 4511 de la nomenclature des installations classées pour la protection de l'environnement et présents dans un entrepôt couvert soumis au régime de l'enregistrement ou de l'autorisation au titre de la rubrique 1510 de cette même nomenclature ",
-        "nor": "DEVP1102086A",
-        "classements": [
-            {"regime": "A", "rubrique": "1436"},
-            {"regime": "A", "rubrique": "4330"},
-            {"regime": "A", "rubrique": "4331"},
-            {"regime": "A", "rubrique": "4510"},
-            {"regime": "A", "rubrique": "4511"},
-            {"regime": "A", "rubrique": "4722"},
-            {"regime": "A", "rubrique": "4734"},
-            {"regime": "A", "rubrique": "4742"},
-            {"regime": "A", "rubrique": "4743"},
-            {"regime": "A", "rubrique": "4744"},
-            {"regime": "A", "rubrique": "4746"},
-            {"regime": "A", "rubrique": "4747"},
-            {"regime": "A", "rubrique": "4748"},
-            {"regime": "A", "rubrique": "1432"},
-        ],
-        "cid": "JORFTEXT000026251890",
-        "state": "ABROGE",
-        "date_of_signature": 1342389600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "43608",
-        "title": "Arrêté du 24/09/20 relatif au stockage en récipients mobiles de liquides inflammables, exploités au sein d'une installation classée pour la protection de l'environnement soumise à autorisation",
-        "nor": "TREP2021860A",
-        "classements": [
-            {"regime": "A", "rubrique": "1436"},
-            {"regime": "A", "rubrique": "4330"},
-            {"regime": "A", "rubrique": "4331"},
-            {"regime": "A", "rubrique": "4510"},
-            {"regime": "A", "rubrique": "4511"},
-            {"regime": "A", "rubrique": "4722"},
-            {"regime": "A", "rubrique": "4734"},
-            {"regime": "A", "rubrique": "4742"},
-            {"regime": "A", "rubrique": "4743"},
-            {"regime": "A", "rubrique": "4744"},
-            {"regime": "A", "rubrique": "4746"},
-            {"regime": "A", "rubrique": "4747"},
-            {"regime": "A", "rubrique": "4748"},
-            {"regime": "A", "rubrique": "1432"},
-        ],
-        "cid": "JORFTEXT000042364605",
-        "state": "VIGUEUR",
-        "date_of_signature": 1600898400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35807",
-        "title": " Arrêté du 18/04/08 relatif aux réservoirs enterrés de liquides inflammables ou combustibles et à leurs équipements annexes exploités au sein d'une installation classée soumise à autorisation, à enregistrement ou à déclaration au titre de l'une ou plusieurs des rubriques n° 1436, 4330, 4331, 4722, 4734, 4742, 4743, 4744, 4746, 4747 ou 4748, ou pour le pétrole brut au titre de l'une ou plusieurs des rubriques n° 4510 ou 4511 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP0804223A",
-        "classements": [
-            {"regime": "A", "rubrique": "1436"},
-            {"regime": "A", "rubrique": "4330"},
-            {"regime": "A", "rubrique": "4331"},
-            {"regime": "A", "rubrique": "4510"},
-            {"regime": "A", "rubrique": "4511"},
-            {"regime": "A", "rubrique": "4722"},
-            {"regime": "A", "rubrique": "4734"},
-            {"regime": "A", "rubrique": "4742"},
-            {"regime": "A", "rubrique": "4743"},
-            {"regime": "A", "rubrique": "4744"},
-            {"regime": "A", "rubrique": "4746"},
-            {"regime": "A", "rubrique": "4747"},
-            {"regime": "A", "rubrique": "4748"},
-        ],
-        "cid": "JORFTEXT000018820571",
-        "state": "VIGUEUR",
-        "date_of_signature": 1208469600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "38104",
-        "title": " Arrêté du 17/08/16 relatif à la prévention des sinistres dans les entrepôts couverts soumis à autorisation sous la rubrique 1510, y compris ceux relevant également de l'une ou plusieurs des rubriques 1530, 1532, 2662 ou 2663 de la nomenclature des installations classées pour la protection de l'environnement (Abrogé) ",
-        "nor": "DEVP1610015A",
-        "classements": [
-            {"regime": "A", "rubrique": "1510"},
-            {"regime": "A", "rubrique": "1530"},
-            {"regime": "A", "rubrique": "1532"},
-            {"regime": "A", "rubrique": "2662"},
-            {"regime": "A", "rubrique": "2663"},
-            {"regime": "E", "rubrique": "1532"},
-        ],
-        "cid": "JORFTEXT000033117556",
-        "state": "ABROGE",
-        "date_of_signature": 1471384800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "39061",
-        "title": " Arrêté du 11/04/17 relatif aux prescriptions générales applicables aux entrepôts couverts soumis à la rubrique 1510, y compris lorsqu'ils relèvent également de l'une ou plusieurs des rubriques 1530, 1532, 2662 ou 2663 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1706393A",
-        "classements": [
-            {"regime": "A", "rubrique": "1510"},
-            {"regime": "E", "rubrique": "1510"},
-            {"regime": "D", "rubrique": "1510"},
-        ],
-        "cid": "JORFTEXT000034429274",
-        "state": "VIGUEUR",
-        "date_of_signature": 1491861600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4445",
-        "title": " Arrêté du 29/09/08 relatif à la prévention des sinistres dans les dépôts de papier et de carton soumis à autorisation au titre de la rubrique n° 1530 de la nomenclature des installations classées ",
-        "nor": "DEVP0823140A",
-        "classements": [{"regime": "A", "rubrique": "1530"}],
-        "cid": "JORFTEXT000019860739",
-        "state": "VIGUEUR",
-        "date_of_signature": 1222639200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35697",
-        "title": " Arrêté du 23/06/15 relatif aux installations mettant en oeuvre des substances radioactives, déchets radioactifs ou résidus solides de minerai d’uranium, de thorium ou de radium soumises à autorisation au titre de la rubrique 1716, de la rubrique 1735 et de la rubrique 2797 de la nomenclature des installations classées ",
-        "nor": "DEVP1425767A",
-        "classements": [
-            {"regime": "A", "rubrique": "1716"},
-            {"regime": "A", "rubrique": "1735"},
-            {"regime": "A", "rubrique": "2797"},
-        ],
-        "cid": "JORFTEXT000030885472",
-        "state": "VIGUEUR",
-        "date_of_signature": 1435010400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28552",
-        "title": " Arrêté du 27/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'autorisation au titre des rubriques n° 2101, 2102, 2111 et 3660 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1329742A",
-        "classements": [
-            {"regime": "A", "rubrique": "2101"},
-            {"regime": "A", "rubrique": "2102"},
-            {"regime": "A", "rubrique": "2111"},
-            {"regime": "A", "rubrique": "3660"},
-        ],
-        "cid": "JORFTEXT000028409297",
-        "state": "VIGUEUR",
-        "date_of_signature": 1388098800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4889",
-        "title": " Arrêté du 31/10/06 fixant les règles techniques auxquelles doivent satisfaire les élevages de lapins soumis à autorisation au titre du livre V du code de l'environnement ",
-        "nor": "DEVP0630181A",
-        "classements": [{"regime": "A", "rubrique": "2110"}],
-        "cid": "JORFTEXT000000246408",
-        "state": "VIGUEUR",
-        "date_of_signature": 1162249200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5129",
-        "title": "Arrêté du 10/08/04 fixant les conditions d'autorisation de détention d'animaux de certaines espèces non domestiques dans les établissements d'élevage, de vente, de location, de transit ou de présentation au public d'animaux d'espèces non domestiques",
-        "nor": "DEVN0430298A",
-        "classements": [{"regime": "A", "rubrique": "2113"}],
-        "cid": "JORFTEXT000000624760",
-        "state": "VIGUEUR",
-        "date_of_signature": 1092088800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "37312",
-        "title": " Arrêté du 15/09/86 fixant les règles techniques auxquelles doivent satisfaire au titre de la protection de l'environnement les élevages d'animaux carnassiers à fourrure ",
-        "classements": [{"regime": "A", "rubrique": "2113"}],
-        "cid": "LEGITEXT000006074852",
-        "state": "VIGUEUR",
-        "date_of_signature": 527119200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4861",
-        "title": " Arrêté du 08/12/06 fixant les règles techniques auxquelles doivent satisfaire les installations renfermant des chiens soumises à autorisation au titre du livre V du code de l'environnement ",
-        "nor": "DEVP0700015A",
-        "classements": [{"regime": "A", "rubrique": "2120"}],
-        "cid": "JORFTEXT000000615876",
-        "state": "VIGUEUR",
-        "date_of_signature": 1165532400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4543",
-        "title": " Arrêté du 01/04/08 fixant les règles techniques auxquelles doivent satisfaire les piscicultures d'eau douce soumises à autorisation au titre du livre V du code de l'environnement (rubrique 2130 de la nomenclature des installations classées) ",
-        "nor": "DEVP0802292A",
-        "classements": [{"regime": "A", "rubrique": "2130"}],
-        "cid": "JORFTEXT000018622600",
-        "state": "VIGUEUR",
-        "date_of_signature": 1207000800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5165",
-        "title": " Arrêté du 25/03/04 fixant les règles générales de fonctionnement et les caractéristiques générales des installations des établissements zoologiques à caractère fixe et permanent, présentant au public des spécimens vivants de la faune locale ou étrangère ",
-        "nor": "DEVN0430016A",
-        "classements": [{"regime": "A", "rubrique": "2140"}],
-        "cid": "JORFTEXT000000610915",
-        "state": "VIGUEUR",
-        "date_of_signature": 1080169200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "39942",
-        "title": "Arrêté du 21/11/17 relatif aux prescriptions générales applicables aux installations relevant du régime de l’autorisation au titre de la rubrique n° 2150 de la nomenclature des installations classées pour la protection de l’environnement",
-        "nor": "TREP1712882A",
-        "cid": "JORFTEXT000036074614",
-        "classements": [{"regime": "A", "rubrique": "2150"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1511218800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5163",
-        "title": " Arrêté du 29/03/04 relatif à la prévention des risques présentés par les silos de céréales, de grains, de produits alimentaires ou de tout autre produit organique dégageant des poussières inflammables ",
-        "nor": "DESP0430052A",
-        "classements": [{"regime": "A", "rubrique": "2160"}],
-        "cid": "JORFTEXT000000797999",
-        "state": "VIGUEUR",
-        "date_of_signature": 1080511200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5161",
-        "title": " Arrêté du 30/04/04 relatif aux prescriptions applicables aux installations classées pour la protection de l'environnement soumises à autorisation sous « les rubriques n° 2210 et 3641 » ",
-        "nor": "DEVP0430124A",
-        "classements": [{"regime": "A", "rubrique": "2210"}],
-        "cid": "JORFTEXT000000803657",
-        "state": "VIGUEUR",
-        "date_of_signature": 1083276000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5491",
-        "title": " Arrêté du 03/05/00 relatif aux prescriptions applicables aux Installations classées pour la protection de l'environnement soumises à autorisation sous la rubrique 2251 (Préparation, conditionnement de vin, la capacité de production étant supérieure à 20 000 hl/an) ",
-        "nor": "ATEP0090178A",
-        "classements": [{"regime": "A", "rubrique": "2251"}],
-        "cid": "JORFTEXT000000399672",
-        "state": "VIGUEUR",
-        "date_of_signature": 957304800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4141",
-        "title": " Arrêté du 18/02/10 relatif à la prévention des risques accidentels présentés par certaines installations classées pour la protection de l'environnement soumises à autorisation sous la rubrique n° 2260 \" broyage, concassage, criblage, déchiquetage, ensachage, pulvérisation, trituration, granulation, nettoyage, tamisage, blutage, mélange, épluchage et décortication des substances végétales et de tous produits organiques naturels, y compris la fabrication d'aliments composés pour animaux (Applicable au régime de l'enregistrement des installations existantes) ",
-        "nor": "DEVP1000790A",
-        "classements": [{"regime": "A", "rubrique": "2260"}],
-        "cid": "JORFTEXT000022039323",
-        "state": "VIGUEUR",
-        "date_of_signature": 1266447600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5509",
-        "title": " Arrêté du 03/04/00 relatif à l'industrie papetière (Abrogé à compter du 1er janvier 2021 - Régime de l'autorisation supprimé au 24/11/2017 pour la rubrique 2440) ",
-        "nor": "ATEP0090163A",
-        "classements": [
-            {"regime": "A", "rubrique": "2430"},
-            {"regime": "A", "rubrique": "2440"},
-            {"regime": "A", "rubrique": "3610"},
-        ],
-        "cid": "JORFTEXT000000570621",
-        "state": "ABROGE",
-        "date_of_signature": 954712800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "43534",
-        "title": " Arrêté du 10/09/20 relatif aux prescriptions générales applicables aux installations relevant du régime de l'autorisation au titre des rubriques 2430 (Préparation de la pâte à papier à l'exclusion des activités classées au titre de la rubrique 3610a), 3610a (Fabrication, dans des installations industrielles, de pâte à papier à partir du bois ou d'autres matières fibreuses) et 3610b (Fabrication, dans des installations industrielles, de papier ou carton, avec une capacité de production supérieure à 20 tonnes par jour) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP2013116A",
-        "classements": [{"regime": "A", "rubrique": "2430"}, {"regime": "A", "rubrique": "3610"}],
-        "cid": "JORFTEXT000042344843",
-        "state": "VIGUEUR",
-        "date_of_signature": 1599688800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5809",
-        "title": " Arrêté du 22/09/94 relatif aux exploitations de carrières ",
-        "nor": "ENVP9430348A",
-        "classements": [{"regime": "A", "rubrique": "2510"}],
-        "cid": "JORFTEXT000000347845",
-        "state": "VIGUEUR",
-        "date_of_signature": 780184800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5859",
-        "title": " Arrêté du 03/05/93 relatif aux cimenteries ",
-        "nor": "ENVP9320194A",
-        "classements": [{"regime": "A", "rubrique": "2520"}],
-        "cid": "JORFTEXT000000346445",
-        "state": "VIGUEUR",
-        "date_of_signature": 736380000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5235",
-        "title": " Arrêté du 12/03/03 relatif à l'industrie du verre et de la fibre minérale ",
-        "nor": "DEVP0320122A",
-        "classements": [{"regime": "A", "rubrique": "2530"}, {"regime": "A", "rubrique": "2531"}],
-        "cid": "JORFTEXT000000411785",
-        "state": "VIGUEUR",
-        "date_of_signature": 1047423600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4919",
-        "title": " Arrêté du 30/06/06 relatif aux « prescriptions générales applicables aux installations relevant du régime de l'autorisation au titre de la rubrique n° 3260 de la nomenclature des installations classées pour la protection de l'environnement » ",
-        "nor": "DEVP0650366A",
-        "classements": [{"regime": "A", "rubrique": "2565"}, {"regime": "A", "rubrique": "3260"}],
-        "cid": "JORFTEXT000000644060",
-        "state": "VIGUEUR",
-        "date_of_signature": 1151618400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5625",
-        "title": " Arrêté du 02/06/98 relatif aux règles techniques auxquelles doivent satisfaire les installations soumises à autorisation au titre de la rubrique 2680-2 de la Nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "ATEP9870206A",
-        "classements": [{"regime": "A", "rubrique": "2680"}],
-        "cid": "JORFTEXT000000389425",
-        "state": "VIGUEUR",
-        "date_of_signature": 896738400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4097",
-        "title": " Arrêté du 19/04/10 relatif à la gestion des déchets des industries extractives ",
-        "nor": "DEVP1010260A",
-        "classements": [{"regime": "A", "rubrique": "2720"}],
-        "cid": "JORFTEXT000022673160",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271628000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5245",
-        "title": " Arrêté du 12/02/03 relatif aux prescriptions applicables aux installations classées soumises à autorisation sous la rubrique 2730 (traitement des cadavres, des déchets ou des sous-produits d'origine animale à l'exclusion des activités visées par d'autres rubriques de la nomenclature) ",
-        "nor": "DEVP0320051A",
-        "classements": [{"regime": "A", "rubrique": "2730"}],
-        "cid": "JORFTEXT000000419523",
-        "state": "VIGUEUR",
-        "date_of_signature": 1045004400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5247",
-        "title": " Arrêté du 12/02/03 relatif aux prescriptions applicables aux installations classées soumises à autorisation sous la rubrique 2731 « de la nomenclature des installations classées pour la protection de l'environnement » ",
-        "nor": "DEVP0320052A",
-        "classements": [{"regime": "A", "rubrique": "2731"}],
-        "cid": "JORFTEXT000000785128",
-        "state": "VIGUEUR",
-        "date_of_signature": 1045004400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4279",
-        "title": " Arrêté du 17/07/09 relatif aux prescriptions applicables aux installations classées soumises à autorisation sous la rubrique 2740 (incinération de cadavres d'animaux de compagnie) (applicable jusqu'au 30 juin 2019 aux installations existantes sauf dispositions mentionnées à l'article 1er de l'arrêté du 06/06/18) ",
-        "nor": "DEVP0913553A",
-        "classements": [{"regime": "A", "rubrique": "2740"}],
-        "cid": "JORFTEXT000021187144",
-        "state": "ABROGE",
-        "date_of_signature": 1247781600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40816",
-        "title": " Arrêté du 06/06/18 relatif aux prescriptions générales applicables aux installations relevant du régime de l'autorisation au titre de la rubrique n° 2740 de la nomenclature des installations classées pour la protection de l'environnement (incinération de cadavres d'animaux) (applicable à compter du 1er juillet 2018 pour les installations nouvelles) ",
-        "nor": "TREP1808485A",
-        "classements": [{"regime": "A", "rubrique": "2740"}],
-        "cid": "JORFTEXT000037078514",
-        "state": "VIGUEUR",
-        "date_of_signature": 1528236000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "37314",
-        "title": " Arrêté du 15/02/16 relatif aux installations de stockage de déchets non dangereux ",
-        "nor": "DEVP1519168A",
-        "classements": [{"regime": "A", "rubrique": "2760"}],
-        "cid": "JORFTEXT000032275960",
-        "state": "VIGUEUR",
-        "date_of_signature": 1455490800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5275",
-        "title": " Arrêté du 20/09/02 relatif aux installations d'incinération et de co-incinération de déchets non dangereux et aux installations incinérant des déchets d'activités de soins à risques infectieux ",
-        "nor": "DEVP0210351A",
-        "classements": [{"regime": "A", "rubrique": "2771"}],
-        "cid": "JORFTEXT000000234557",
-        "state": "VIGUEUR",
-        "date_of_signature": 1032472800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4529",
-        "title": " Arrêté du 22/04/08 fixant les règles techniques auxquelles doivent satisfaire les installations de compostage soumises à autorisation en application du titre Ier du livre V du code de l'environnement ",
-        "nor": "DEVP0810090A",
-        "classements": [{"regime": "A", "rubrique": "2780"}],
-        "cid": "JORFTEXT000018800981",
-        "state": "VIGUEUR",
-        "date_of_signature": 1208815200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4223",
-        "title": " Arrêté du 10/11/09 fixant les règles techniques auxquelles doivent satisfaire les installations de méthanisation soumises à autorisation en application du titre Ier du livre V du code de l'environnement ",
-        "nor": "DEVP0920874A",
-        "classements": [{"regime": "A", "rubrique": "2781"}],
-        "cid": "JORFTEXT000021334497",
-        "state": "VIGUEUR",
-        "date_of_signature": 1257807600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "27198",
-        "title": " Arrêté du 26/08/13 relatif aux installations de combustion d'une puissance supérieure ou égale à 20 MW soumises à autorisation au titre de la rubrique 2910 et de la rubrique 2931 ",
-        "nor": "DEVP1300515A",
-        "classements": [{"regime": "A", "rubrique": "2910"}, {"regime": "A", "rubrique": "2931"}],
-        "cid": "JORFTEXT000027995928",
-        "state": "ABROGE",
-        "date_of_signature": 1377468000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5199",
-        "title": " Arrêté du 30/07/03 relatif aux chaudières présentes dans des installations existantes de combustion d'une puissance supérieure à 20 MWth (abrogé) ",
-        "nor": "DEVP0320297A",
-        "classements": [{"regime": "A", "rubrique": "2910"}],
-        "cid": "JORFTEXT000000248169",
-        "state": "ABROGE",
-        "date_of_signature": 1059516000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5297",
-        "title": " Arrêté du 20/06/02 relatif aux chaudières présentes dans une installation nouvelle ou modifiée d'une puissance supérieure à 20 MWth (abrogé) ",
-        "nor": "DEVP0210222A",
-        "classements": [{"regime": "A", "rubrique": "2910"}],
-        "cid": "JORFTEXT000000779198",
-        "state": "ABROGE",
-        "date_of_signature": 1024524000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5551",
-        "title": " Arrêté du 11/08/99 relatif à la réduction des émissions polluantes des moteurs et turbines à combustion ainsi que des chaudières utilisées en postcombustion soumis à autorisation sous la rubrique 2910 de la nomenclature des installations classées pour la protection de l'environnement (abrogé) ",
-        "nor": "ATEP9980359A",
-        "classements": [{"regime": "A", "rubrique": "2910"}],
-        "cid": "JORFTEXT000000763103",
-        "state": "ABROGE",
-        "date_of_signature": 934322400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41041",
-        "title": " Arrêté du 03/08/18 relatif aux installations de combustion d'une puissance thermique nominale totale inférieure à 50 MW soumises à autorisation au titre des rubriques 2910, 2931 ou 3110 (applicable à compter du 20 décembre 2018) ",
-        "nor": "TREP1726534A",
-        "classements": [{"regime": "A", "rubrique": "2910"}, {"regime": "A", "rubrique": "3110"}],
-        "cid": "JORFTEXT000037284792",
-        "state": "VIGUEUR",
-        "date_of_signature": 1533247200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "37522",
-        "nor": "DEVP1525037A",
-        "cid": "JORFTEXT000032575814",
-        "title": "Arrêté du 23/05/16 relatif aux installations de production de chaleur et/ou d'électricité à partir de déchets non dangereux préparés sous forme de combustibles solides de récupération dans des installations prévues à cet effet associés ou non à un autre combustible et relevant de la rubrique 2971 de la nomenclature des installations classées pour la protection de l'environnement",
-        "classements": [{"regime": "A", "rubrique": "2971"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1463954400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35876",
-        "title": " Arrêté du 26/08/11 relatif aux installations de production d’électricité utilisant l’énergie mécanique du vent au sein d’une installation soumise à autorisation au titre de la rubrique 2980 de la législation des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1119348A",
-        "classements": [{"regime": "A", "rubrique": "2980"}],
-        "cid": "JORFTEXT000024507365",
-        "state": "VIGUEUR",
-        "date_of_signature": 1314309600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41057",
-        "title": " Arrêté du 03/08/18 relatif aux installations de combustion d'une puissance thermique nominale totale supérieure ou égale à 50 MW soumises à autorisation au titre de la rubrique 3110 (applicable à compter du 20 décembre 2018) ",
-        "nor": "TREP1726535A",
-        "classements": [{"regime": "A", "rubrique": "3110"}],
-        "cid": "JORFTEXT000037284870",
-        "state": "VIGUEUR",
-        "date_of_signature": 1533247200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35378",
-        "title": " Arrêté du 06/11/07 relatif à la prévention des risques présentés par les dépôts et ateliers utilisant des « substances ou mélanges relevant de l'une ou plusieurs des rubriques nos 4410, 4411, 4420, 4421 ou 4422 » ",
-        "nor": "DEVP0765936A",
-        "classements": [
-            {"regime": "A", "rubrique": "1212"},
-            {"regime": "A", "rubrique": "4410"},
-            {"regime": "A", "rubrique": "4411"},
-            {"regime": "A", "rubrique": "4420"},
-            {"regime": "A", "rubrique": "4421"},
-            {"regime": "A", "rubrique": "4422"},
-        ],
-        "cid": "JORFTEXT000017734021",
-        "state": "VIGUEUR",
-        "date_of_signature": 1194303600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35408",
-        "title": " Arrêté du 13/04/10 relatif à la prévention des risques présentés par les stockages d'engrais solides à base de nitrate d'ammonium soumis à autorisation au titre de la rubrique 4702 et les stockages de produits soumis à autorisation au titre de la rubrique 4703 ",
-        "nor": "DEVP1008014A",
-        "classements": [
-            {"regime": "A", "rubrique": "4702"},
-            {"regime": "A", "rubrique": "4703"},
-            {"regime": "A", "rubrique": "1331"},
-            {"regime": "A", "rubrique": "1332"},
-        ],
-        "cid": "JORFTEXT000022092771",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271109600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35412",
-        "title": " Arrêté du 02/01/08 relatif aux réservoirs fixes manufacturés de gaz inflammables liquéfiés, de capacité unitaire supérieure ou égale à 50 tonnes, présents au sein d'une installation classée pour la protection de l'environnement relevant du régime de l'autorisation au titre de la rubrique n° 4718 de la nomenclature des installations classées, à l'exception des stockages réfrigérés ou cryogéniques ",
-        "nor": "DEVP0773312A",
-        "classements": [{"regime": "A", "rubrique": "4718"}, {"regime": "A", "rubrique": "1412"}],
-        "cid": "JORFTEXT000018008144",
-        "state": "VIGUEUR",
-        "date_of_signature": 1199228400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35821",
-        "title": " Arrêté du 08/12/95 relatif à la lutte contre les émissions de composés organiques volatils résultant du stockage de l'essence et de sa distribution des terminaux aux stations-service ",
-        "nor": "ENVP9540375A",
-        "classements": [{"regime": "A", "rubrique": "4734"}, {"regime": "A", "rubrique": "1434", "alinea": "1"}],
-        "cid": "JORFTEXT000000190990",
-        "state": "VIGUEUR",
-        "date_of_signature": 818377200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5687",
-        "title": " Arrêté du 16/07/97 relatif aux installations de réfrigération employant l'ammoniac comme fluide frigorigène soumises à autorisation au titre de la rubrique n° 4735 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "ATEP9760318A",
-        "classements": [{"regime": "A", "rubrique": "4735"}],
-        "cid": "JORFTEXT000000202155",
-        "state": "VIGUEUR",
-        "date_of_signature": 869004000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4121",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux stations-service relevant du régime de l'enregistrement au titre de la rubrique n° 1435 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1001977A",
-        "classements": [{"regime": "E", "rubrique": "1435"}],
-        "cid": "JORFTEXT000022097882",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4117",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux entrepôts couverts relevant du régime de l'enregistrement au titre de la rubrique n° 1510 de la nomenclature des installations classées pour la protection de l'environnement (Abrogé) ",
-        "nor": "DEVP1001986A",
-        "classements": [{"regime": "E", "rubrique": "1510"}],
-        "cid": "JORFTEXT000022097930",
-        "state": "ABROGE",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4119",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux entrepôts frigorifiques relevant du régime de l'enregistrement au titre de la rubrique n° 1511 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1001990A",
-        "classements": [{"regime": "E", "rubrique": "1511"}],
-        "cid": "JORFTEXT000022097955",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4115",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux dépôts de papier et de carton relevant du régime de l'enregistrement au titre de la rubrique n° 1530 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1001980A",
-        "classements": [{"regime": "E", "rubrique": "1530"}],
-        "cid": "JORFTEXT000022097909",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "27126",
-        "title": " Arrêté du 11/09/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 1532 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1243055A",
-        "classements": [{"regime": "E", "rubrique": "1532"}],
-        "cid": "JORFTEXT000027945320",
-        "state": "VIGUEUR",
-        "date_of_signature": 1378850400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28554",
-        "title": " Arrêté du 27/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre des rubriques n° « 2101 », 2102 et 2111 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1329749A",
-        "classements": [
-            {"regime": "E", "rubrique": "2101"},
-            {"regime": "E", "rubrique": "2102"},
-            {"regime": "E", "rubrique": "2111"},
-        ],
-        "cid": "JORFTEXT000028409384",
-        "state": "VIGUEUR",
-        "date_of_signature": 1388098800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41355",
-        "title": " Arrêté du 22/10/18 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2120 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1815609A",
-        "classements": [{"regime": "E", "rubrique": "2120"}],
-        "cid": "JORFTEXT000037518992",
-        "state": "VIGUEUR",
-        "date_of_signature": 1540159200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "22363",
-        "title": " Arrêté du 26/11/12 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2160 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1235599A",
-        "classements": [{"regime": "E", "rubrique": "2160"}],
-        "cid": "JORFTEXT000026694666",
-        "state": "VIGUEUR",
-        "date_of_signature": 1353884400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28562",
-        "title": " Arrêté du 14/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2220 (préparation ou conservation de produits alimentaires d’origine végétale) de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1307960A",
-        "classements": [{"regime": "E", "rubrique": "2220"}],
-        "cid": "JORFTEXT000028374558",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386975600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3501",
-        "title": " Arrêté du 23/03/12 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2221 (préparation ou conservation de produits alimentaires d’origine animale) de la nomenclature des installations classées pour la protection de l’environnement  ",
-        "nor": "DEVP1205541A",
-        "classements": [{"regime": "E", "rubrique": "2221"}],
-        "cid": "JORFTEXT000025673104",
-        "state": "VIGUEUR",
-        "date_of_signature": 1332457200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "39127",
-        "title": "Arrêté du 24/04/17 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2230 de la nomenclature des installations classées pour la protection de l’environnement",
-        "nor": "DEVP1634272A",
-        "cid": "JORFTEXT000034485050",
-        "classements": [{"regime": "E", "rubrique": "2230"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1492984800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "39125",
-        "title": "Arrêté du 24/04/17 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2240 de la nomenclature des installations classées pour la protection de l'environnement",
-        "nor": "DEVP1628601A",
-        "cid": "JORFTEXT000034484917",
-        "classements": [{"regime": "E", "rubrique": "2240"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1492984800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3885",
-        "title": " Arrêté du 14/01/11 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2250 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1023820A",
-        "classements": [{"regime": "E", "rubrique": "2250"}],
-        "cid": "JORFTEXT000023689944",
-        "state": "VIGUEUR",
-        "date_of_signature": 1294959600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "22368",
-        "title": " Arrêté du 26/11/12 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2251 (préparation, conditionnement de vins) de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1236050A",
-        "classements": [{"regime": "E", "rubrique": "2251"}],
-        "cid": "JORFTEXT000026694774",
-        "state": "VIGUEUR",
-        "date_of_signature": 1353884400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41357",
-        "title": " Arrêté du 22/10/18 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2260 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1815737A",
-        "classements": [{"regime": "E", "rubrique": "2260"}],
-        "cid": "JORFTEXT000037519057",
-        "state": "VIGUEUR",
-        "date_of_signature": 1540159200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3883",
-        "title": " Arrêté du 14/01/11 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2340 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1029977A",
-        "classements": [{"regime": "E", "rubrique": "2340"}],
-        "cid": "JORFTEXT000023713720",
-        "state": "VIGUEUR",
-        "date_of_signature": 1294959600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "32625",
-        "title": " Arrêté du 02/09/14 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2410 (installation où l'on travaille le bois ou matériaux combustibles analogues) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1403460A",
-        "classements": [{"regime": "E", "rubrique": "2410"}],
-        "cid": "JORFTEXT000029419186",
-        "state": "VIGUEUR",
-        "date_of_signature": 1409608800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "22367",
-        "title": " Arrêté du 26/11/12 relatif aux prescriptions générales applicables aux installations de broyage, concassage, criblage, etc., relevant du régime de l'enregistrement au titre de la rubrique n° 2515 de la nomenclature des installations classées pour la protection de l'environnement «, y compris lorsqu'elles relèvent également de l'une ou plusieurs des rubriques n° 2516 ou 2517 » ",
-        "nor": "DEVP1235896A",
-        "classements": [{"regime": "E", "rubrique": "2515"}],
-        "cid": "JORFTEXT000026694913",
-        "state": "VIGUEUR",
-        "date_of_signature": 1353884400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28564",
-        "title": " Arrêté du 10/12/13 relatif aux prescriptions générales applicables aux stations de transit de produits minéraux pulvérulents non ensachés ou de déchets non dangereux inertes pulvérulents relevant du régime de l'enregistrement au titre de la rubrique n° 2516 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1329349A",
-        "classements": [{"regime": "E", "rubrique": "2516"}],
-        "cid": "JORFTEXT000028379475",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386630000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28570",
-        "title": " Arrêté du 10/12/13 relatif aux prescriptions générales applicables aux stations de transit de produits minéraux ou de déchets non dangereux inertes autres que ceux visés par d'autres rubriques relevant du régime de l'enregistrement au titre de la rubrique n° 2517 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1329353A",
-        "classements": [{"regime": "E", "rubrique": "2517"}],
-        "cid": "JORFTEXT000028379599",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386630000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3723",
-        "title": " Arrêté du 08/08/11 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2518 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1102170A",
-        "classements": [{"regime": "E", "rubrique": "2518"}],
-        "cid": "JORFTEXT000024490187",
-        "state": "VIGUEUR",
-        "date_of_signature": 1312754400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41901",
-        "title": " Arrêté du 09/04/19 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2521 de la nomenclature des installations classées pour la protection de l'environnement - Enrobage au bitume de matériaux routiers (Centrale d') ",
-        "nor": "TREP1900331A",
-        "classements": [{"regime": "E", "rubrique": "2521"}],
-        "cid": "JORFTEXT000038358856",
-        "state": "VIGUEUR",
-        "date_of_signature": 1554760800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3721",
-        "title": " Arrêté du 08/08/11 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2522 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1102168A",
-        "classements": [{"regime": "E", "rubrique": "2522"}],
-        "cid": "JORFTEXT000024490314",
-        "state": "VIGUEUR",
-        "date_of_signature": 1312754400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28575",
-        "title": " Arrêté du 14/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2560 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1326230A",
-        "classements": [{"regime": "E", "rubrique": "2560"}],
-        "cid": "JORFTEXT000028374698",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386975600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28590",
-        "title": " Arrêté du 14/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2563 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1326229A",
-        "classements": [{"regime": "E", "rubrique": "2563"}],
-        "cid": "JORFTEXT000028374816",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386975600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41897",
-        "title": " Arrêté du 09/04/19 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2564 (nettoyage, dégraissage, décapage de surfaces par des procédés utilisant des liquides organohalogénés ou des solvants organiques) ou de la rubrique n° 2565 (revêtement métallique ou traitement de surfaces par voie électrolytique ou chimique) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1835514A",
-        "classements": [{"regime": "E", "rubrique": "2564"}, {"regime": "E", "rubrique": "2565"}],
-        "cid": "JORFTEXT000038358667",
-        "state": "VIGUEUR",
-        "date_of_signature": 1554760800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28599",
-        "title": " Arrêté du 27/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2661 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1326202A",
-        "classements": [{"regime": "E", "rubrique": "2661"}],
-        "cid": "JORFTEXT000028409140",
-        "state": "VIGUEUR",
-        "date_of_signature": 1388098800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4107",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux stockages de polymères (matières plastiques, caoutchoucs, élastomères, résines et adhésifs synthétiques) relevant du régime de l'enregistrement au titre de la rubrique n° 2662 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1001998A",
-        "classements": [{"regime": "E", "rubrique": "2662"}],
-        "cid": "JORFTEXT000022198234",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4105",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux stockages de pneumatiques et de produits dont 50 % au moins de la masse totale unitaire est composée de polymères (matières plastiques, caoutchouc, élastomères, résines et adhésifs synthétiques) relevant du régime de l'enregistrement au titre de la rubrique n° 2663 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1002002A",
-        "classements": [{"regime": "E", "rubrique": "2663"}],
-        "cid": "JORFTEXT000022198256",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "33655",
-        "title": " Arrêté du 12/12/14 relatif aux prescriptions générales applicables aux installations du régime de l'enregistrement relevant de la rubrique n° 2760 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1412526A",
-        "classements": [{"regime": "E", "rubrique": "2760"}],
-        "cid": "JORFTEXT000029893853",
-        "state": "VIGUEUR",
-        "date_of_signature": 1418338800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3465",
-        "title": " Arrêté du 20/04/12 relatif aux prescriptions générales applicables aux installations classées de compostage soumises à enregistrement sous la rubrique n° 2780 ",
-        "nor": "DEVP1221724A",
-        "classements": [{"regime": "E", "rubrique": "2780"}],
-        "cid": "JORFTEXT000025789288",
-        "state": "VIGUEUR",
-        "date_of_signature": 1334872800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40755",
-        "title": "Arrêté du 06/06/18 relatif aux prescriptions générales applicables aux installations de broyage de déchets végétaux non dangereux relevant du régime de l'enregistrement au titre de la rubrique n° 2794 de la nomenclature des installations classées pour la protection de l'environnement",
-        "nor": "TREP1800790A",
-        "cid": "JORFTEXT000037032291",
-        "classements": [{"regime": "E", "rubrique": "2794"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1528236000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40749",
-        "title": "Arrêté du 18/05/18 relatif aux prescriptions générales applicables aux installations de broyage de déchets végétaux non dangereux relevant du régime de la déclaration au titre de la rubrique n° 2794 de la nomenclature des installations classées pour la protection de l'environnement",
-        "nor": "TREP1800788A",
-        "cid": "JORFTEXT000036976496",
-        "classements": [{"regime": "D", "rubrique": "2794"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1526594400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28582",
-        "title": " Arrêté du 14/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2921 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1305353A",
-        "classements": [{"regime": "E", "rubrique": "2921"}],
-        "cid": "JORFTEXT000028374941",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386975600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35386",
-        "title": " Arrêté du 29/07/10 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 4220 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1018696A",
-        "classements": [{"regime": "E", "rubrique": "4220"}, {"regime": "E", "rubrique": "1311"}],
-        "cid": "JORFTEXT000022593820",
-        "state": "VIGUEUR",
-        "date_of_signature": 1280354400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35344",
-        "title": " Arrêté du 01/06/15 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de l‘une au moins des rubriques 4331 ou 4734 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1501359A",
-        "classements": [{"regime": "E", "rubrique": "4331"}, {"regime": "E", "rubrique": "4734"}],
-        "cid": "JORFTEXT000030673177",
-        "state": "VIGUEUR",
-        "date_of_signature": 1433109600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5647",
-        "title": " Arrêté du 23/02/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1136 : Emploi ou stockage de l'ammoniac ",
-        "nor": "ATEP9870095A",
-        "classements": [{"regime": "D", "rubrique": "1136", "state": "SUPPRIMEE"}],
-        "cid": "JORFTEXT000000753815",
-        "state": "ABROGE",
-        "date_of_signature": 888188400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5671",
-        "title": " Arrêté du 08/10/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1139 : Fabrication, stockage ou emploi du dioxyde de chlore ",
-        "nor": "ATEP9760448A",
-        "classements": [{"regime": "D", "rubrique": "1139"}],
-        "cid": "JORFTEXT000000370053",
-        "state": "ABROGE",
-        "date_of_signature": 876261600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5501",
-        "title": " Arrêté du 10/04/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1141 (Chlorure d'hydrogène anhydre (emploi ou stockage du) ",
-        "nor": "ATEP0090165A",
-        "classements": [{"regime": "D", "rubrique": "1141"}],
-        "cid": "JORFTEXT000000765521",
-        "state": "ABROGE",
-        "date_of_signature": 955317600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5319",
-        "title": " Arrêté du 02/05/02 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1155  \" Agro-pharmaceutiques \" (dépôts de produits), à l'exclusion des substances et préparations visées par les rubriques n° 1111 et n° 1150, et des liquides inflammables de catégorie A au sens de la rubrique n° 1430 ",
-        "nor": "ATEP0210164A",
-        "classements": [{"regime": "D", "rubrique": "1155"}],
-        "cid": "JORFTEXT000000777935",
-        "state": "ABROGE",
-        "date_of_signature": 1020290400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5503",
-        "title": " Arrêté du 10/04/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1156 \" oxydes d'azote autres que l'hemioxyde d'azote \" (emploi ou stockage des) ",
-        "nor": "ATEP0090166A",
-        "classements": [{"regime": "D", "rubrique": "1156"}],
-        "cid": "JORFTEXT000000400526",
-        "state": "ABROGE",
-        "date_of_signature": 955317600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4665",
-        "title": " Arrêté du 29/10/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1158 (Emploi ou stockage de diisocyanate de diphénylméthane) ",
-        "nor": "DEVP0770386A",
-        "classements": [{"regime": "D", "rubrique": "1158"}],
-        "cid": "JORFTEXT000017640584",
-        "state": "VIGUEUR",
-        "date_of_signature": 1193612400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5333",
-        "title": " Arrêté du 02/04/02 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1185 ",
-        "nor": "ATEP0210151A",
-        "classements": [{"regime": "D", "rubrique": "1185"}],
-        "cid": "JORFTEXT000000592074",
-        "state": "ABROGE",
-        "date_of_signature": 1017698400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "32401",
-        "title": " Arrêté du 04/08/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4802 (Rubrique devenue la rubrique 1185 à compter du 25 octobre 2018) ",
-        "nor": "DEVP1402942A",
-        "classements": [{"regime": "D", "rubrique": "1185"}, {"regime": "D", "rubrique": "4802"}],
-        "cid": "JORFTEXT000029359291",
-        "state": "VIGUEUR",
-        "date_of_signature": 1407103200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35414",
-        "title": " Arrêté du 07/01/03 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous l'une ou plusieurs des rubriques nos 1413 ou 4718 de la nomenclature des installations classées ",
-        "nor": "DEVP0320022A",
-        "classements": [{"regime": "D", "rubrique": "1413"}, {"regime": "D", "rubrique": "4718"}],
-        "cid": "JORFTEXT000000418677",
-        "state": "VIGUEUR",
-        "date_of_signature": 1041894000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41364",
-        "title": " Arrêté du 22/10/18 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1416 (station de distribution d'hydrogène gazeux) de la nomenclature des installations classées et modifiant l'arrêté du 26 novembre 2015 relatif aux prescriptions générales applicables aux installations mettant en œuvre l'hydrogène gazeux dans une installation classée pour la protection de l'environnement pour alimenter des chariots à hydrogène gazeux lorsque la quantité d'hydrogène présente au sein de l'établissement relève du régime de la déclaration pour la rubrique n° 4715 et modifiant l'arrêté du 4 août 2014 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4802 (applicable à compter du 1er janvier 2019) ",
-        "nor": "TREP1816561A",
-        "classements": [{"regime": "D", "rubrique": "1416"}],
-        "cid": "JORFTEXT000037519292",
-        "state": "VIGUEUR",
-        "date_of_signature": 1540159200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35846",
-        "title": " Arrêté du 15/04/10 relatif aux prescriptions générales applicables aux stations-service soumises à déclaration sous la rubrique n° 1435 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1001974A",
-        "classements": [{"regime": "D", "rubrique": "1435"}],
-        "cid": "JORFTEXT000022097853",
-        "state": "VIGUEUR",
-        "date_of_signature": 1271282400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35811",
-        "title": " Arrêté du 22/12/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous l'une ou plusieurs des rubriques nos 1436, 4330, 4331, 4722, 4734, 4742, 4743, 4744, 4746, 4747 ou 4748, ou pour le pétrole brut sous l'une ou plusieurs des rubriques nos 4510 ou 4511 ",
-        "nor": "DEVP0827876A",
-        "classements": [
-            {"regime": "D", "rubrique": "1436"},
-            {"regime": "D", "rubrique": "4330"},
-            {"regime": "D", "rubrique": "4331"},
-            {"regime": "D", "rubrique": "4510"},
-            {"regime": "D", "rubrique": "4511"},
-            {"regime": "D", "rubrique": "4722"},
-            {"regime": "D", "rubrique": "4734"},
-            {"regime": "D", "rubrique": "4742"},
-            {"regime": "D", "rubrique": "4743"},
-            {"regime": "D", "rubrique": "4744"},
-            {"regime": "D", "rubrique": "4746"},
-            {"regime": "D", "rubrique": "4747"},
-            {"regime": "D", "rubrique": "4748"},
-            {"regime": "D", "rubrique": "1432"},
-        ],
-        "cid": "JORFTEXT000019996765",
-        "state": "VIGUEUR",
-        "date_of_signature": 1229900400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35813",
-        "title": " Arrêté du 20/04/05 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous l'une ou plusieurs des rubriques nos 1436, 4330, 4331, 4722, 4734, 4742, 4743, 4744, 4746, 4747 ou 4748, ou pour le pétrole brut sous l'une ou plusieurs des rubriques nos 4510 ou 4511 ",
-        "nor": "DEVP0540156A",
-        "classements": [
-            {"regime": "D", "rubrique": "1436"},
-            {"regime": "D", "rubrique": "4330"},
-            {"regime": "D", "rubrique": "4331"},
-            {"regime": "D", "rubrique": "4510"},
-            {"regime": "D", "rubrique": "4511"},
-            {"regime": "D", "rubrique": "4722"},
-            {"regime": "D", "rubrique": "4734"},
-            {"regime": "D", "rubrique": "4742"},
-            {"regime": "D", "rubrique": "4743"},
-            {"regime": "D", "rubrique": "4744"},
-            {"regime": "D", "rubrique": "4746"},
-            {"regime": "D", "rubrique": "4747"},
-            {"regime": "D", "rubrique": "1433"},
-        ],
-        "cid": "JORFTEXT000000630475",
-        "state": "VIGUEUR",
-        "date_of_signature": 1113948000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "38417",
-        "title": " Arrêté du 05/12/16 relatif aux prescriptions applicables à certaines installations classées pour la protection de l’environnement soumises à déclaration (applicable jusqu'au 31/12/19 pour ce qui concerne les rubriques 4440, 4441 ou 4442) ",
-        "nor": "DEVP1628687A",
-        "classements": [
-            {"regime": "D", "rubrique": "1450"},
-            {"regime": "D", "rubrique": "1532"},
-            {"regime": "D", "rubrique": "2113"},
-            {"regime": "D", "rubrique": "2171"},
-            {"regime": "D", "rubrique": "2175"},
-            {"regime": "D", "rubrique": "2180"},
-            {"regime": "D", "rubrique": "2230"},
-            {"regime": "D", "rubrique": "2240"},
-            {"regime": "D", "rubrique": "2252"},
-            {"regime": "D", "rubrique": "2275"},
-            {"regime": "D", "rubrique": "2311"},
-            {"regime": "D", "rubrique": "2321"},
-            {"regime": "D", "rubrique": "2350"},
-            {"regime": "D", "rubrique": "2355"},
-            {"regime": "D", "rubrique": "2410"},
-            {"regime": "D", "rubrique": "2420"},
-            {"regime": "D", "rubrique": "2430"},
-            {"regime": "D", "rubrique": "2440"},
-            {"regime": "D", "rubrique": "2445"},
-            {"regime": "D", "rubrique": "2546"},
-            {"regime": "D", "rubrique": "2630"},
-            {"regime": "D", "rubrique": "2631"},
-            {"regime": "D", "rubrique": "2640"},
-            {"regime": "D", "rubrique": "2690"},
-            {"regime": "D", "rubrique": "2915"},
-            {"regime": "D", "rubrique": "4320"},
-            {"regime": "D", "rubrique": "4321"},
-            {"regime": "D", "rubrique": "4442"},
-            {"regime": "D", "rubrique": "4705"},
-            {"regime": "D", "rubrique": "4706"},
-            {"regime": "D", "rubrique": "4716"},
-            {"regime": "D", "rubrique": "4801"},
-            {"regime": "D", "rubrique": "1414", "alinea": "2"},
-            {"regime": "D", "rubrique": "2130", "alinea": "2"},
-        ],
-        "cid": "JORFTEXT000033560858",
-        "state": "VIGUEUR",
-        "date_of_signature": 1480892400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5489",
-        "title": " Arrêté du 03/05/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1455  \" Carbure de calcium (stockage) \" ",
-        "nor": "ATEP0090187A",
-        "classements": [{"regime": "D", "rubrique": "1455"}],
-        "cid": "JORFTEXT000000399730",
-        "state": "VIGUEUR",
-        "date_of_signature": 957304800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4375",
-        "title": " Arrêté du 23/12/08 relatif aux prescriptions générales applicables aux entrepôts couverts relevant du régime de la déclaration au titre de la rubrique n° 1510 de la nomenclature des installations classées pour la protection de l'environnement (Abrogé) ",
-        "nor": "DEVP0827962A",
-        "classements": [{"regime": "D", "rubrique": "1510"}],
-        "cid": "JORFTEXT000020009023",
-        "state": "ABROGE",
-        "date_of_signature": 1229986800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "30277",
-        "title": " Arrêté du 27/03/14 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1511 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1405235A",
-        "classements": [{"regime": "D", "rubrique": "1511"}],
-        "cid": "JORFTEXT000028814820",
-        "state": "VIGUEUR",
-        "date_of_signature": 1395874800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4443",
-        "title": " Arrêté du 30/09/08 relatif aux prescriptions générales applicables aux dépôts de papier et carton relevant du régime de la déclaration au titre de la rubrique n° 1530 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP0823142A",
-        "classements": [{"regime": "D", "rubrique": "1530"}],
-        "cid": "JORFTEXT000019831346",
-        "state": "VIGUEUR",
-        "date_of_signature": 1222725600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5507",
-        "title": " Arrêté du 03/04/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1531, \" Stockages, par voie humide (immersion ou aspersion) de bois non traité chimiquement \" ",
-        "nor": "ATEP0090149A",
-        "classements": [{"regime": "D", "rubrique": "1531"}],
-        "cid": "JORFTEXT000000764048",
-        "state": "VIGUEUR",
-        "date_of_signature": 954712800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5467",
-        "title": " Arrêté du 06/09/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1611 ",
-        "nor": "ATEP0090353A",
-        "classements": [{"regime": "D", "rubrique": "1611"}],
-        "cid": "JORFTEXT000000573432",
-        "state": "VIGUEUR",
-        "date_of_signature": 968191200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4659",
-        "title": " Arrêté du 31/10/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1612 (Emploi ou stockage d'acide chlorosulfurique ou d'oléums) ",
-        "nor": "DEVP0770381A",
-        "classements": [{"regime": "D", "rubrique": "1612"}],
-        "cid": "JORFTEXT000017640613",
-        "state": "VIGUEUR",
-        "date_of_signature": 1193785200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5397",
-        "title": " Arrêté du 26/07/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 1630 ",
-        "nor": "ATEP0100249A",
-        "classements": [{"regime": "D", "rubrique": "1630"}],
-        "cid": "JORFTEXT000000772915",
-        "state": "VIGUEUR",
-        "date_of_signature": 996098400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "33687",
-        "title": " Arrêté du 03/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique 1716-2 ",
-        "nor": "DEVP1414981A",
-        "classements": [{"regime": "D", "rubrique": "1716"}],
-        "cid": "JORFTEXT000029917414",
-        "state": "VIGUEUR",
-        "date_of_signature": 1417561200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "42852",
-        "title": " Arrêté du 13/12/19 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration au titre de la rubrique n° 1978 (installations et activités utilisant des solvants organiques) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1935133A",
-        "classements": [{"regime": "D", "rubrique": "1978"}],
-        "cid": "JORFTEXT000041413405",
-        "state": "VIGUEUR",
-        "date_of_signature": 1576191600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5091",
-        "title": " Arrêté du 07/02/05 fixant les règles techniques auxquelles doivent satisfaire les élevages de bovins, de volailles et/ou de gibier à plumes et de porcs soumis à déclaration au titre du livre V du code de l’environnement ",
-        "nor": "DEVP0540078A",
-        "classements": [
-            {"regime": "D", "rubrique": "2101"},
-            {"regime": "D", "rubrique": "2102"},
-            {"regime": "D", "rubrique": "2111"},
-        ],
-        "cid": "JORFTEXT000000444709",
-        "state": "ABROGE",
-        "date_of_signature": 1107730800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28556",
-        "title": " Arrêté du 27/12/13 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous les rubriques n° 2101-1, 2101-2, 2101-3, 2102 et 2111 ",
-        "nor": "DEVP1329745A",
-        "classements": [
-            {"regime": "D", "rubrique": "2102"},
-            {"regime": "D", "rubrique": "2111"},
-            {"regime": "D", "rubrique": "2101", "alinea": "1"},
-            {"regime": "D", "rubrique": "2101", "alinea": "2"},
-            {"regime": "D", "rubrique": "2101", "alinea": "3"},
-        ],
-        "cid": "JORFTEXT000028409472",
-        "state": "VIGUEUR",
-        "date_of_signature": 1388098800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4891",
-        "title": " Arrêté du 30/10/06 fixant les règles techniques auxquelles doivent satisfaire les élevages de lapins soumis à déclaration au titre du livre V du code de l'environnement ",
-        "nor": "DEVP0630182A",
-        "classements": [{"regime": "D", "rubrique": "2110"}],
-        "cid": "JORFTEXT000000425804",
-        "state": "VIGUEUR",
-        "date_of_signature": 1162162800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5083",
-        "title": " Arrêté du 10/02/05 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2112 (activité « couvoirs ») ",
-        "nor": "DEVP0540081A",
-        "classements": [{"regime": "D", "rubrique": "2112"}],
-        "cid": "JORFTEXT000000444667",
-        "state": "VIGUEUR",
-        "date_of_signature": 1107990000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4863",
-        "title": " Arrêté du 08/12/06 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2120 ",
-        "nor": "DEVP0700016A",
-        "classements": [{"regime": "D", "rubrique": "2120"}],
-        "cid": "JORFTEXT000000791749",
-        "state": "VIGUEUR",
-        "date_of_signature": 1165532400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "39978",
-        "title": " Arrêté du 21/11/17 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration avec contrôle périodique au titre de la rubrique n° 2150 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1712920A",
-        "classements": [{"regime": "D", "rubrique": "2150"}],
-        "cid": "JORFTEXT000036074702",
-        "state": "VIGUEUR",
-        "date_of_signature": 1511218800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5359",
-        "title": " Arrêté du 07/01/02 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2170 \" engrais et supports de culture (fabrication des) à partir de matières organiques \" et mettant en œuvre un procédé de transformation biologique aérobie (compostage) des matières organiques (Abrogé) ",
-        "nor": "ATEP0210030A",
-        "classements": [{"regime": "D", "rubrique": "2170"}],
-        "cid": "JORFTEXT000000593036",
-        "state": "ABROGE",
-        "date_of_signature": 1010358000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5159",
-        "title": " Arrêté du 30/04/04 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2210 « abattage d'animaux » ",
-        "nor": "DEVP0430125A",
-        "classements": [{"regime": "D", "rubrique": "2210"}],
-        "cid": "JORFTEXT000000418135",
-        "state": "VIGUEUR",
-        "date_of_signature": 1083276000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5045",
-        "title": " Arrêté du 17/06/05 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2220 Préparation ou conservation de produits alimentaires d'origine végétale par cuisson, appertisation, surgélation, congélation, lyophilisation, déshydratation, torréfaction, etc., à l'exclusion du sucre, de la fécule, du malt, des huiles et des aliments pour le bétail mais y compris les ateliers de maturation de fruits et légumes ",
-        "nor": "DEVP0540254A",
-        "classements": [{"regime": "D", "rubrique": "2220"}],
-        "cid": "JORFTEXT000000812308",
-        "state": "VIGUEUR",
-        "date_of_signature": 1118959200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4727",
-        "title": " Arrêté du 09/08/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2221 ",
-        "nor": "DEVP0763217A",
-        "classements": [{"regime": "D", "rubrique": "2221"}],
-        "cid": "JORFTEXT000000277959",
-        "state": "VIGUEUR",
-        "date_of_signature": 1186610400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3427",
-        "title": " Arrêté du 25/05/12 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2250 (production par distillation d'alcools de bouche d'origine agricole)  ",
-        "nor": "DEVP1220106A",
-        "classements": [{"regime": "D", "rubrique": "2250"}],
-        "cid": "JORFTEXT000026140269",
-        "state": "VIGUEUR",
-        "date_of_signature": 1337896800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5571",
-        "title": "Arrêté du 15/03/99 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2251 (Préparation, conditionnement de vin, la capacité de production étant supérieure à 500 hl/an mais inférieure ou égale à 20 000 hl/an).",
-        "nor": "ATEP9980125A",
-        "cid": "JORFTEXT000000210657",
-        "classements": [{"regime": "D", "rubrique": "2251"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 921452400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "34787",
-        "title": " Arrêté du 18/12/14 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration au titre de la rubrique n° 2253 (préparation, conditionnement de boissons, bière, jus de fruits, autres boissons, à l'exclusion des eaux minérales, eaux de source, eaux de table et des activités visées par les rubriques 2230, 2250, 2251 et 2252) de la nomenclature des installations classées pour la protection de l'environnement (Rubrique supprimée à compter du 25 octobre 2018) ",
-        "nor": "DEVP1421693A",
-        "classements": [{"regime": "D", "rubrique": "2253"}],
-        "cid": "JORFTEXT000030426541",
-        "state": "ABROGE",
-        "date_of_signature": 1418857200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4935",
-        "title": " Arrêté du 23/05/06 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2260 « broyage, concassage, criblage, déchiquetage, ensachage, pulvérisation, trituration, nettoyage, tamisage, blutage, mélange, épluchage et décortication des substances végétales et de tous produits organiques naturels, à l'exclusion des activités visées par les rubriques nos 2220, 2221, 2225 et 2226, mais y compris la fabrication d'aliments pour le bétail » ",
-        "nor": "DEVP0650343A",
-        "classements": [{"regime": "D", "rubrique": "2260"}],
-        "cid": "JORFTEXT000000792853",
-        "state": "VIGUEUR",
-        "date_of_signature": 1148335200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "33845",
-        "title": " Arrêté du 18/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2265 ",
-        "nor": "DEVP1422570A",
-        "classements": [{"regime": "D", "rubrique": "2265"}],
-        "cid": "JORFTEXT000030024830",
-        "state": "VIGUEUR",
-        "date_of_signature": 1418857200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5403",
-        "title": " Arrêté du 25/07/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2330 « Teinture, impression, apprêt, enduction, blanchiment et délavage de matières textiles » ",
-        "nor": "ATEP0100260A",
-        "cid": "JORFTEXT000000582313",
-        "classements": [{"regime": "D", "rubrique": "2330"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 996012000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3887",
-        "title": " Arrêté du 14/01/11 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2340 ",
-        "nor": "DEVP1029979A",
-        "classements": [{"regime": "D", "rubrique": "2340"}],
-        "cid": "JORFTEXT000023596449",
-        "state": "VIGUEUR",
-        "date_of_signature": 1294959600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5321",
-        "title": " Arrêté du 02/05/02 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2345 relative à l'utilisation de solvants pour le nettoyage à sec et le traitement des textiles ou des vêtements ",
-        "nor": "ATEP0210158A",
-        "classements": [{"regime": "D", "rubrique": "2345"}],
-        "cid": "JORFTEXT000000413046",
-        "state": "ABROGE",
-        "date_of_signature": 1020290400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4259",
-        "title": " Arrêté du 31/08/09 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2345 relative à l'utilisation de solvants pour le nettoyage à sec et le traitement des textiles ou des vêtements ",
-        "nor": "DEVP0915322A",
-        "classements": [{"regime": "D", "rubrique": "2345"}],
-        "cid": "JORFTEXT000021030687",
-        "state": "VIGUEUR",
-        "date_of_signature": 1251669600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5401",
-        "title": " Arrêté du 25/07/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2351 « Teinture et pigmentation de peaux » ",
-        "nor": "ATEP0100261A",
-        "cid": "JORFTEXT000000215138",
-        "classements": [{"regime": "D", "rubrique": "2351"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 996012000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5399",
-        "title": " Arrêté du 25/07/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2360 « Ateliers de fabrication de chaussures, maroquinerie ou travail des cuirs et des peaux » ",
-        "nor": "ATEP0100262A",
-        "cid": "JORFTEXT000000582314",
-        "classements": [{"regime": "D", "rubrique": "2360"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 996012000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5105",
-        "title": " Arrêté du 17/12/04 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2415 relative aux installations de mise en œuvre de produits de préservation du bois et matériaux dérivés ",
-        "nor": "DEVP0540024A",
-        "classements": [{"regime": "D", "rubrique": "2415"}],
-        "cid": "JORFTEXT000000786106",
-        "state": "VIGUEUR",
-        "date_of_signature": 1103238000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5207",
-        "title": " Arrêté du 16/07/03 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2450 relative aux imprimeries ou ateliers de reproduction graphique sur tout support tel que métal, papier, carton, matières plastiques, textiles, etc., utilisant une forme imprimante ",
-        "nor": "DEVP0320218A",
-        "classements": [{"regime": "D", "rubrique": "2450"}],
-        "cid": "JORFTEXT000000432338",
-        "state": "VIGUEUR",
-        "date_of_signature": 1058306400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4851",
-        "title": " Arrêté du 26/12/06 relatif aux prescriptions générales applicables aux exploitations de carrières soumises à déclaration sous la rubrique n° 2510 de la nomenclature des installations classées ",
-        "nor": "DEVP0700014A",
-        "classements": [{"regime": "D", "rubrique": "2510"}],
-        "cid": "JORFTEXT000000427983",
-        "state": "VIGUEUR",
-        "date_of_signature": 1167087600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5689",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'Environnement soumises à déclaration sous la rubrique n° 2515 : \" Broyage, concassage, criblage, ensachage, pulvérisation, nettoyage, tamisage, mélange de pierres, cailloux, minerais et autres produits minéraux naturels ou artificiels \" ",
-        "nor": "ATEP9760290A",
-        "classements": [{"regime": "D", "rubrique": "2515"}],
-        "cid": "JORFTEXT000000552021",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5691",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2516 : \" Station de transit de produits minéraux pulvérulents non ensachés tels que ciments, plâtres, chaux, sables fillerisés \" ",
-        "nor": "ATEP9760291A",
-        "classements": [{"regime": "D", "rubrique": "2516"}],
-        "cid": "JORFTEXT000000369331",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5693",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2517 : \" Station de transit de produits minéraux solides à l'exclusion de ceux visés par d'autres rubriques \" ",
-        "nor": "ATEP9760292A",
-        "classements": [{"regime": "D", "rubrique": "2517"}],
-        "cid": "JORFTEXT000000369330",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3641",
-        "title": " Arrêté du 26/11/11 relatif aux prescriptions générales applicables aux installations de fabrication de béton prêt à l’emploi, soumises à déclaration sous la rubrique n° 2518 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1103455A",
-        "classements": [{"regime": "D", "rubrique": "2518"}],
-        "cid": "JORFTEXT000024968167",
-        "state": "VIGUEUR",
-        "date_of_signature": 1322262000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5695",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2521 : \" Enrobage au bitume de matériaux routiers (centrales) à froid \" ",
-        "nor": "ATEP9760293A",
-        "classements": [{"regime": "D", "rubrique": "2521"}],
-        "cid": "JORFTEXT000000552020",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5697",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2522 : \" Matériel vibrant (emploi de) pour la fabrication de matériaux tels que béton, agglomérés,etc,\"(Abrogé) ",
-        "nor": "ATEP9760294A",
-        "classements": [{"regime": "D", "rubrique": "2522"}],
-        "cid": "JORFTEXT000000552019",
-        "state": "ABROGE",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3639",
-        "title": " Arrêté du 26/11/11 relatif aux prescriptions générales applicables aux installations de fabrication de produits en béton, soumises à déclaration sous la rubrique n° 2522 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1103457A",
-        "classements": [{"regime": "D", "rubrique": "2522"}],
-        "cid": "JORFTEXT000024968178",
-        "state": "VIGUEUR",
-        "date_of_signature": 1322262000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5699",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2524 : \" Minéraux naturels ou artificiels tels que le granit, l'ardoise, le verre, etc (Ateliers de taillage, sciage et polissage de) \" ",
-        "nor": "ATEP9760295A",
-        "classements": [{"regime": "D", "rubrique": "2524"}],
-        "cid": "JORFTEXT000000369301",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4825",
-        "title": " Arrêté du 14/02/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2530 relative à la fabrication et au travail du verre ",
-        "nor": "DEVP0700197A",
-        "classements": [{"regime": "D", "rubrique": "2530"}],
-        "cid": "JORFTEXT000000793557",
-        "state": "VIGUEUR",
-        "date_of_signature": 1171407600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4823",
-        "title": " Arrêté du 14/02/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2531 relative au travail chimique du verre ",
-        "nor": "DEVP0700198A",
-        "classements": [{"regime": "D", "rubrique": "2531"}],
-        "cid": "JORFTEXT000000793558",
-        "state": "VIGUEUR",
-        "date_of_signature": 1171407600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5701",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2550 : \" Fonderie (fabrication de produits moulés) \" de plomb et alliages contenant du plomb (au moins 3%) ",
-        "nor": "ATEP9760296A",
-        "classements": [{"regime": "D", "rubrique": "2550"}],
-        "cid": "JORFTEXT000000734395",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5703",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2551 : \" Fonderie (fabrication de produits moulés) \" de métaux et alliages ferreux ",
-        "nor": "ATEP9760297A",
-        "classements": [{"regime": "D", "rubrique": "2551"}],
-        "cid": "JORFTEXT000000182387",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5705",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2552 : Fonderie (fabrication de produits moulés) de métaux et alliages non ferreux (à l’exception de celles relevant de la rubrique n° 2550) ",
-        "nor": "ATEP9760298A",
-        "classements": [{"regime": "D", "rubrique": "2552"}],
-        "cid": "JORFTEXT000000552051",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35912",
-        "title": " Arrêté du 27/07/15 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2560 : applicable au 1er janvier 2016 ",
-        "nor": "DEVP1510020A",
-        "classements": [{"regime": "D", "rubrique": "2560"}],
-        "cid": "JORFTEXT000031027065",
-        "state": "VIGUEUR",
-        "date_of_signature": 1437948000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5707",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2560 : \" Métaux et alliages (travail mécanique des) \" : Abrogé à compter du 31 décembre 2015 ",
-        "nor": "ATEP9760299A",
-        "cid": "JORFTEXT000000183180",
-        "classements": [{"regime": "D", "rubrique": "2560"}],
-        "state": "ABROGE",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "36024",
-        "title": " Arrêté du 27/07/15 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2561 : applicable au 1er janvier 2016 ",
-        "nor": "DEVP1510050A",
-        "classements": [{"regime": "D", "rubrique": "2561"}],
-        "cid": "JORFTEXT000031027083",
-        "state": "VIGUEUR",
-        "date_of_signature": 1437948000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5709",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2561 : \" Métaux et alliages (trempe, recuit ou revenu) \" : abrogé à compter du 1er janvier 2016 ",
-        "nor": "ATEP9760300A",
-        "cid": "JORFTEXT000000183179",
-        "classements": [{"regime": "D", "rubrique": "2561"}],
-        "state": "ABROGE",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5711",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2562 : \" Bains de sels fondus(chauffage et traitements industriels par l’intermédiaire de) \" ",
-        "nor": "ATEP9760301A",
-        "classements": [{"regime": "D", "rubrique": "2562"}],
-        "cid": "JORFTEXT000000369300",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35916",
-        "title": " Arrêté du 27/07/15 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2563 : applicable au 1er janvier 2016 ",
-        "nor": "DEVP1510052A",
-        "classements": [{"regime": "D", "rubrique": "2563"}],
-        "cid": "JORFTEXT000031027101",
-        "state": "VIGUEUR",
-        "date_of_signature": 1437948000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5145",
-        "title": " Arrêté du 21/06/04 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique 2564 relative au nettoyage, dégraissage, décapage de surfaces (métaux, matières plastiques...) par des procédés utilisant des liquides organohalogénés ou des solvants organiques (Abrogé) ",
-        "nor": "DEVP0430194A",
-        "classements": [{"regime": "D", "rubrique": "2564"}],
-        "cid": "JORFTEXT000000619925",
-        "state": "ABROGE",
-        "date_of_signature": 1087768800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41899",
-        "title": " Arrêté du 09/04/19 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration avec contrôle périodique sous la rubrique n° 2564 (nettoyage, dégraissage, décapage de surfaces par des procédés utilisant des liquides organohalogénés ou des solvants organiques) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1835517A",
-        "classements": [{"regime": "D", "rubrique": "2564"}],
-        "cid": "JORFTEXT000038358780",
-        "state": "VIGUEUR",
-        "date_of_signature": 1554760800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5713",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2565 : Métaux et matières plastiques (traitement des) pour le dégraissage, le décapage, la conversion, le polissage, la métallisation, etc., par voie électrolytique, chimique, ou par emploi de liquides halogénés ",
-        "nor": "ATEP9760302A",
-        "classements": [{"regime": "D", "rubrique": "2565"}],
-        "cid": "JORFTEXT000000551978",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "36022",
-        "title": " Arrêté du 27/07/15 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2566 : applicable au 1er janvier 2016 ",
-        "nor": "DEVP1430922A",
-        "classements": [{"regime": "D", "rubrique": "2566"}],
-        "cid": "JORFTEXT000031027118",
-        "state": "VIGUEUR",
-        "date_of_signature": 1437948000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "36020",
-        "title": " Arrêté du 27/07/15 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2567 : applicable au 1er janvier 2016 ",
-        "nor": "DEVP1430916A",
-        "classements": [{"regime": "D", "rubrique": "2567"}],
-        "cid": "JORFTEXT000031027133",
-        "state": "VIGUEUR",
-        "date_of_signature": 1437948000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4293",
-        "title": " Arrêté du 07/07/09 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2570 ",
-        "nor": "DEVP0915020A",
-        "classements": [{"regime": "D", "rubrique": "2570"}],
-        "cid": "JORFTEXT000021028173",
-        "state": "VIGUEUR",
-        "date_of_signature": 1246917600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5715",
-        "title": " Arrêté du 30/06/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2575 : \" Abrasives (emploi de matières) telles que sables, corindon, grenailles métalliques, etc., sur un matériau quelconque pour gravure, dépolissage, décapage, grainage \" ",
-        "nor": "ATEP9760303A",
-        "classements": [{"regime": "D", "rubrique": "2575"}],
-        "cid": "JORFTEXT000000369299",
-        "state": "VIGUEUR",
-        "date_of_signature": 867621600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5521",
-        "title": " Arrêté du 14/01/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2660 ou 2661 (Fabrication, régénération ou transformation de polymères [matières plastiques, caoutchouc, élastomères, résines et adhésifs synthétiques]) ",
-        "nor": "ATEP0090027A",
-        "classements": [{"regime": "D", "rubrique": "2660"}, {"regime": "D", "rubrique": "2661"}],
-        "cid": "JORFTEXT000000202785",
-        "state": "VIGUEUR",
-        "date_of_signature": 947804400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5519",
-        "title": " Arrêté du 14/01/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2662 (Stockage de polymères [matières plastiques, caoutchouc, élastomères, résines et adhésifs synthétiques) ",
-        "nor": "ATEP0090028A",
-        "classements": [{"regime": "D", "rubrique": "2662"}],
-        "cid": "JORFTEXT000000202754",
-        "state": "VIGUEUR",
-        "date_of_signature": 947804400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5517",
-        "title": " Arrêté du 14/01/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2663 (Stockage de pneumatiques et de produits dont 50 % au moins de la masse totale unitaire est composée de polymères [matières plastiques, caoutchouc, élastomères, résines et adhésifs synthétiques) ",
-        "nor": "ATEP0090029A",
-        "classements": [{"regime": "D", "rubrique": "2663"}],
-        "cid": "JORFTEXT000000567521",
-        "state": "VIGUEUR",
-        "date_of_signature": 947804400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5125",
-        "title": " Arrêté du 13/10/04 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2685 « Médicaments (fabrication et division en vue de la préparation de) à usage humain ou vétérinaire » (caduc) ",
-        "nor": "DEVP0430382A",
-        "classements": [{"regime": "D", "rubrique": "2685"}],
-        "cid": "JORFTEXT000000807264",
-        "state": "VIGUEUR",
-        "date_of_signature": 1097618400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5725",
-        "title": " Arrêté du 02/04/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2710 : \" Déchetteries aménagées pour la collecte des encombrants, matériaux ou produits triés et apportés par le public \" (Abrogé) ",
-        "nor": "ENVP9760118A",
-        "classements": [{"regime": "D", "rubrique": "2710"}],
-        "cid": "JORFTEXT000000565555",
-        "state": "ABROGE",
-        "date_of_signature": 859932000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4633",
-        "title": " Arrêté du 12/12/07 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2711 « Transit, regroupement, tri, désassemblage, remise en état d'équipements électriques et électroniques mis au rebut »  (Abrogé) ",
-        "nor": "DEVP0772484A",
-        "classements": [{"regime": "D", "rubrique": "2711"}],
-        "cid": "JORFTEXT000017883576",
-        "state": "ABROGE",
-        "date_of_signature": 1197414000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40992",
-        "title": " Arrêté du 06/06/18 relatif aux prescriptions générales applicables aux installations de transit, regroupement, tri ou préparation en vue de la réutilisation de déchets relevant du régime de la déclaration au titre de la rubrique n° 2711 (déchets d'équipements électriques et électroniques), 2713 (métaux ou déchets de métaux non dangereux, alliage de métaux ou déchets d'alliage de métaux non dangereux), 2714 (déchets non dangereux de papiers, cartons, plastiques, caoutchouc, textiles, bois) ou 2716 (déchets non dangereux non inertes) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1800782A",
-        "classements": [
-            {"regime": "D", "rubrique": "2711"},
-            {"regime": "D", "rubrique": "2713"},
-            {"regime": "D", "rubrique": "2714"},
-            {"regime": "D", "rubrique": "2716"},
-        ],
-        "cid": "JORFTEXT000037032196",
-        "state": "VIGUEUR",
-        "date_of_signature": 1528236000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3985",
-        "title": " Arrêté du 13/10/10 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2713 (Abrogé) ",
-        "nor": "DEVP1022261A",
-        "classements": [{"regime": "D", "rubrique": "2713"}],
-        "cid": "JORFTEXT000023022747",
-        "state": "ABROGE",
-        "date_of_signature": 1286920800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3983",
-        "title": " Arrêté du 14/10/10 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2714 (Abrogé) ",
-        "nor": "DEVP1022264A",
-        "classements": [{"regime": "D", "rubrique": "2714"}],
-        "cid": "JORFTEXT000023022756",
-        "state": "ABROGE",
-        "date_of_signature": 1287007200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3981",
-        "title": " Arrêté du 15/10/10 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 2715 ",
-        "nor": "DEVP1022266A",
-        "classements": [{"regime": "D", "rubrique": "2715"}],
-        "cid": "JORFTEXT000023022769",
-        "state": "VIGUEUR",
-        "date_of_signature": 1287093600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3979",
-        "title": " Arrêté du 16/10/10 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2716 (Abrogé) ",
-        "nor": "DEVP1022267A",
-        "classements": [{"regime": "D", "rubrique": "2716"}],
-        "cid": "JORFTEXT000023022778",
-        "state": "ABROGE",
-        "date_of_signature": 1287180000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3385",
-        "title": " Arrêté du 30/07/12 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2719 (installation temporaire de transit de déchets issus de pollutions accidentelles marines ou fluviales ou de déchets issus de catastrophes naturelles)  ",
-        "nor": "DEVP1230975A",
-        "classements": [{"regime": "D", "rubrique": "2719"}],
-        "cid": "JORFTEXT000026274584",
-        "state": "VIGUEUR",
-        "date_of_signature": 1343599200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3755",
-        "title": " Arrêté du 12/07/11 relatif aux prescriptions générales applicables aux installations classées de compostage soumises à déclaration sous la rubrique n° 2780 ",
-        "nor": "DEVP1117266A",
-        "classements": [{"regime": "D", "rubrique": "2780"}],
-        "cid": "JORFTEXT000024446891",
-        "state": "VIGUEUR",
-        "date_of_signature": 1310421600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3643",
-        "title": " Arrêté du 23/11/11 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique 2791 (installation de traitement de déchets non dangereux à l’exclusion des installations visées aux rubriques 2720, 2760, 2771, 2780, 2781 et 2782) ",
-        "nor": "DEVP1130388A",
-        "classements": [{"regime": "D", "rubrique": "2791"}],
-        "cid": "JORFTEXT000024908356",
-        "state": "VIGUEUR",
-        "date_of_signature": 1322002800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35868",
-        "title": " Arrêté du 23/12/11 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2795 ",
-        "nor": "DEVP1135517A",
-        "classements": [{"regime": "D", "rubrique": "2795"}],
-        "cid": "JORFTEXT000025126196",
-        "state": "VIGUEUR",
-        "date_of_signature": 1324594800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "33685",
-        "title": " Arrêté du 03/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique 2798 ",
-        "nor": "DEVP1414980A",
-        "classements": [{"regime": "D", "rubrique": "2798"}],
-        "cid": "JORFTEXT000029917403",
-        "state": "VIGUEUR",
-        "date_of_signature": 1417561200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5683",
-        "title": " Arrêté du 25/07/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2910 : Combustion (Abrogé) ",
-        "nor": "ATEP9760321A",
-        "classements": [{"regime": "D", "rubrique": "2910"}],
-        "cid": "JORFTEXT000000202304",
-        "state": "ABROGE",
-        "date_of_signature": 869781600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5109",
-        "title": " Arrêté du 13/12/04 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2921 Installations de refroidissement par dispersion d'eau dans un flux d'air (Abrogé) ",
-        "nor": "DEVP0430481A",
-        "classements": [{"regime": "D", "rubrique": "2921"}],
-        "cid": "JORFTEXT000000424501",
-        "state": "ABROGE",
-        "date_of_signature": 1102892400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "28559",
-        "title": " Arrêté du 14/12/13 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration au titre de la rubrique n° 2921 de la nomenclature des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1305345A",
-        "classements": [{"regime": "D", "rubrique": "2921"}],
-        "cid": "JORFTEXT000028375083",
-        "state": "VIGUEUR",
-        "date_of_signature": 1386975600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5481",
-        "title": " Arrêté du 29/05/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2925 “ accumulateurs (ateliers de charge d') ” - (Rubrique n°2925-1) ",
-        "nor": "ATEP0090222A",
-        "classements": [{"regime": "D", "rubrique": "2925"}],
-        "cid": "JORFTEXT000000584145",
-        "state": "VIGUEUR",
-        "date_of_signature": 959551200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41031",
-        "title": " Arrêté du 03/08/18 relatif aux prescriptions générales applicables aux ateliers de charge contenant au moins 10 véhicules de transport en commun de catégorie M2 ou M3 fonctionnant grâce à l’énergie électrique et soumis à déclaration sous la rubrique n° 2925 de la nomenclature des installations classées pour la protection de l’environnement (Rubrique n°2925-2) ",
-        "nor": "TREP1813684A",
-        "classements": [{"regime": "D", "rubrique": "2925"}],
-        "cid": "JORFTEXT000037311523",
-        "state": "VIGUEUR",
-        "date_of_signature": 1533247200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5149",
-        "title": " Arrêté du 04/06/04 relatif aux prescriptions générales applicables aux  installations classées soumises à déclaration sous la rubrique 2930 relative aux ateliers de réparation et d'entretien de véhicules et engins à moteur, y compris les activités de carrosserie et de tôlerie. ",
-        "nor": "DEVP0430152A",
-        "classements": [{"regime": "D", "rubrique": "2930"}],
-        "cid": "JORFTEXT000000619924",
-        "state": "VIGUEUR",
-        "date_of_signature": 1086300000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5505",
-        "title": " Arrêté du 03/04/00 relatif aux prescriptions applicables aux installations classées soumises à déclaration ou à autorisation au titre de la rubrique n° 2935 \" Parcs de stationnement couverts et garages-hôtels de véhicules à moteur \" relatives à l'accès des véhicules utilisant les gaz de pétrole liquéfiés dans leur système de propulsion (Rubrique supprimée à compter du 1er juillet 2006) ",
-        "nor": "ATEP0090148A",
-        "classements": [{"regime": "D", "rubrique": "2935"}],
-        "cid": "JORFTEXT000000764047",
-        "state": "ABROGE",
-        "date_of_signature": 954712800,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5323",
-        "title": " Arrêté du 02/05/02 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique 2940 ",
-        "nor": "ATEP0210160A",
-        "classements": [{"regime": "D", "rubrique": "2940"}],
-        "cid": "JORFTEXT000000413047",
-        "state": "VIGUEUR",
-        "date_of_signature": 1020290400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5735",
-        "title": " Arrêté du 23/01/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2950 : \" Traitement et développement de surfaces photosensibles à base argentique \" ",
-        "nor": "ENVP9760037A",
-        "classements": [{"regime": "D", "rubrique": "2950"}],
-        "cid": "JORFTEXT000000199989",
-        "state": "VIGUEUR",
-        "date_of_signature": 853974000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35874",
-        "title": " Arrêté du 26/08/11 relatif aux installations de production d’électricité utilisant l’énergie mécanique du vent au sein d’une installation soumise à déclaration au titre de la rubrique 2980 de la législation des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1119342A",
-        "classements": [{"regime": "D", "rubrique": "2980"}],
-        "cid": "JORFTEXT000024507356",
-        "state": "VIGUEUR",
-        "date_of_signature": 1314309600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35318",
-        "title": " Arrêté du 13/07/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous l'une ou plusieurs des rubriques nos 4110,4709,4713,4736 ou 4737 ",
-        "nor": "ATEP9870263A",
-        "classements": [
-            {"regime": "D", "rubrique": "4110"},
-            {"regime": "D", "rubrique": "4709"},
-            {"regime": "D", "rubrique": "4713"},
-            {"regime": "D", "rubrique": "4736"},
-            {"regime": "D", "rubrique": "4737"},
-            {"regime": "D", "rubrique": "1111"},
-        ],
-        "cid": "JORFTEXT000000756621",
-        "state": "VIGUEUR",
-        "date_of_signature": 900280800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35330",
-        "title": " Arrêté du 13/07/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous l'une ou plusieurs des rubriques nos 4120, 4130, 4140, 4150, 4738, 4739 ou 4740 ",
-        "nor": "ATEP9870264A",
-        "classements": [
-            {"regime": "D", "rubrique": "4120"},
-            {"regime": "D", "rubrique": "4130"},
-            {"regime": "D", "rubrique": "4140"},
-            {"regime": "D", "rubrique": "4150"},
-            {"regime": "D", "rubrique": "4738"},
-            {"regime": "D", "rubrique": "4739"},
-            {"regime": "D", "rubrique": "4740"},
-            {"regime": "D", "rubrique": "1131"},
-        ],
-        "cid": "JORFTEXT000000756654",
-        "state": "VIGUEUR",
-        "date_of_signature": 900280800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "38030",
-        "title": "Arrêté du 12/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4210",
-        "nor": "DEVP1404789A",
-        "cid": "JORFTEXT000029893770",
-        "classements": [{"regime": "D", "rubrique": "4210"}, {"regime": "D", "rubrique": "1310"}],
-        "state": "VIGUEUR",
-        "date_of_signature": 1418338800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35384",
-        "title": " Arrêté du 29/02/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4220 ",
-        "nor": "DEVP0800186A",
-        "classements": [{"regime": "D", "rubrique": "4220"}, {"regime": "D", "rubrique": "1311"}],
-        "cid": "JORFTEXT000018332514",
-        "state": "VIGUEUR",
-        "date_of_signature": 1204239600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35358",
-        "title": " Arrêté du 10/11/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous « l'une ou plusieurs des rubriques nos 4410, 4411, 4420, 4421 ou 4422 » ",
-        "nor": "DEVP0825190A",
-        "classements": [
-            {"regime": "D", "rubrique": "4410"},
-            {"regime": "D", "rubrique": "4411"},
-            {"regime": "D", "rubrique": "4420"},
-            {"regime": "D", "rubrique": "4421"},
-            {"regime": "D", "rubrique": "4422"},
-            {"regime": "D", "rubrique": "1212"},
-        ],
-        "cid": "JORFTEXT000019776628",
-        "state": "VIGUEUR",
-        "date_of_signature": 1226271600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "42468",
-        "title": " Arrêté du 01/08/19 relatif aux prescriptions générales applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous l’une au moins des rubriques no 4440, 4441 ou 4442 (applicable à compter du 1er janvier 2020) ",
-        "nor": "TREP1915883A",
-        "classements": [
-            {"regime": "D", "rubrique": "4440"},
-            {"regime": "D", "rubrique": "4441"},
-            {"regime": "D", "rubrique": "4442"},
-        ],
-        "cid": "JORFTEXT000039104029",
-        "state": "VIGUEUR",
-        "date_of_signature": 1564610400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35352",
-        "title": " Arrêté du 23/12/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous « l'une ou plusieurs des rubriques nos 4510, 4741 ou 4745 » ",
-        "nor": "ATEP9980022A",
-        "classements": [
-            {"regime": "D", "rubrique": "4510"},
-            {"regime": "D", "rubrique": "4741"},
-            {"regime": "D", "rubrique": "4745"},
-            {"regime": "D", "rubrique": "1172"},
-        ],
-        "cid": "JORFTEXT000000392482",
-        "state": "VIGUEUR",
-        "date_of_signature": 914367600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35354",
-        "title": " Arrêté du 23/12/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4511 ",
-        "nor": "ATEP9980023A",
-        "classements": [{"regime": "D", "rubrique": "4511"}, {"regime": "D", "rubrique": "1173"}],
-        "cid": "JORFTEXT000000574817",
-        "state": "VIGUEUR",
-        "date_of_signature": 914367600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35855",
-        "title": " Arrêté du 15/05/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4610 ",
-        "nor": "ATEP0100162A",
-        "classements": [{"regime": "D", "rubrique": "4610"}, {"regime": "D", "rubrique": "1810"}],
-        "cid": "JORFTEXT000000770271",
-        "state": "VIGUEUR",
-        "date_of_signature": 989877600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35858",
-        "title": " Arrêté du 15/05/01 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous l'une ou plusieurs des rubriques nos 4620 ou 4630 ",
-        "nor": "ATEP0100163A",
-        "classements": [
-            {"regime": "D", "rubrique": "4620"},
-            {"regime": "D", "rubrique": "4630"},
-            {"regime": "D", "rubrique": "1820"},
-        ],
-        "cid": "JORFTEXT000000770272",
-        "state": "VIGUEUR",
-        "date_of_signature": 989877600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35388",
-        "title": " Arrêté du 18/12/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4701 ",
-        "nor": "DEVP0827973A",
-        "classements": [{"regime": "D", "rubrique": "4701"}, {"regime": "D", "rubrique": "1330"}],
-        "cid": "JORFTEXT000019982114",
-        "state": "VIGUEUR",
-        "date_of_signature": 1229554800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35406",
-        "title": " Arrêté du 06/07/06 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4702 ",
-        "nor": "DEVP0650389A",
-        "classements": [{"regime": "D", "rubrique": "4702"}, {"regime": "D", "rubrique": "1331"}],
-        "cid": "JORFTEXT000000609220",
-        "state": "VIGUEUR",
-        "date_of_signature": 1152136800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35348",
-        "title": " Arrêté du 30/10/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous « l'une ou plusieurs des rubriques nos 4707, 4711, 4717, 4723, 4724, 4726, 4728, 4729, 4730, 4732 ou 4733 » ",
-        "nor": "DEVP0770176A",
-        "classements": [
-            {"regime": "D", "rubrique": "4707"},
-            {"regime": "D", "rubrique": "4711"},
-            {"regime": "D", "rubrique": "4717"},
-            {"regime": "D", "rubrique": "4723"},
-            {"regime": "D", "rubrique": "4724"},
-            {"regime": "D", "rubrique": "4726"},
-            {"regime": "D", "rubrique": "4728"},
-            {"regime": "D", "rubrique": "4729"},
-            {"regime": "D", "rubrique": "4730"},
-            {"regime": "D", "rubrique": "4732"},
-            {"regime": "D", "rubrique": "4733"},
-            {"regime": "D", "rubrique": "1150"},
-        ],
-        "cid": "JORFTEXT000017640600",
-        "state": "VIGUEUR",
-        "date_of_signature": 1193698800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35334",
-        "title": " Arrêté du 17/12/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4710 ",
-        "nor": "DEVP0827976A",
-        "classements": [{"regime": "D", "rubrique": "4710"}, {"regime": "D", "rubrique": "1138"}],
-        "cid": "JORFTEXT000019982074",
-        "state": "VIGUEUR",
-        "date_of_signature": 1229468400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35337",
-        "title": " Arrêté du 02/11/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4714 ",
-        "nor": "DEVP0770150A",
-        "classements": [{"regime": "D", "rubrique": "4714"}, {"regime": "D", "rubrique": "1140"}],
-        "cid": "JORFTEXT000017640626",
-        "state": "VIGUEUR",
-        "date_of_signature": 1193958000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35416",
-        "title": " Arrêté du 12/02/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4715 ",
-        "nor": "ATEP9870096A",
-        "classements": [{"regime": "D", "rubrique": "4715"}, {"regime": "D", "rubrique": "1416"}],
-        "cid": "JORFTEXT000000571176",
-        "state": "VIGUEUR",
-        "date_of_signature": 887238000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35410",
-        "title": " Arrêté du 23/08/05 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4718 de la nomenclature des installations classées ",
-        "nor": "DEVP0540337A",
-        "classements": [{"regime": "D", "rubrique": "4718"}, {"regime": "D", "rubrique": "1412"}],
-        "cid": "JORFTEXT000000447400",
-        "state": "VIGUEUR",
-        "date_of_signature": 1124748000,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35801",
-        "title": " Arrêté du 10/03/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'Environnement soumises à déclaration sous la rubrique n° 4719 ",
-        "nor": "ENVP9760090A",
-        "classements": [{"regime": "D", "rubrique": "4719"}, {"regime": "D", "rubrique": "1418"}],
-        "cid": "JORFTEXT000000564940",
-        "state": "VIGUEUR",
-        "date_of_signature": 857948400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35803",
-        "title": " Arrêté du 06/05/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'Environnement soumises à déclaration sous l'une ou plusieurs des rubriques nos 4720 ou 4721 ",
-        "nor": "ENVP9760152A",
-        "classements": [
-            {"regime": "D", "rubrique": "4720"},
-            {"regime": "D", "rubrique": "4721"},
-            {"regime": "D", "rubrique": "1419"},
-        ],
-        "cid": "JORFTEXT000000383619",
-        "state": "VIGUEUR",
-        "date_of_signature": 862869600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35382",
-        "title": " Arrêté du 10/03/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises a déclaration sous la rubrique n° 4725 ",
-        "nor": "ENVP9760089A",
-        "classements": [{"regime": "D", "rubrique": "4725"}, {"regime": "D", "rubrique": "1220"}],
-        "cid": "JORFTEXT000000381759",
-        "state": "VIGUEUR",
-        "date_of_signature": 857948400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35322",
-        "title": " Arrêté du 23/08/05 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4727 ",
-        "nor": "DEVP0540338A",
-        "classements": [{"regime": "D", "rubrique": "4727"}, {"regime": "D", "rubrique": "1116"}],
-        "cid": "JORFTEXT000000446932",
-        "state": "VIGUEUR",
-        "date_of_signature": 1124748000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35350",
-        "title": " Arrêté du 08/08/97 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 4731 ",
-        "nor": "ATEP9760346A",
-        "classements": [{"regime": "D", "rubrique": "4731"}, {"regime": "D", "rubrique": "1157"}],
-        "cid": "JORFTEXT000000551230",
-        "state": "VIGUEUR",
-        "date_of_signature": 870991200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35370",
-        "title": " Arrêté du 19/11/09 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 4735 ",
-        "nor": "DEVP0927337A",
-        "classements": [{"regime": "D", "rubrique": "4735"}, {"regime": "D", "rubrique": "1136"}],
-        "cid": "JORFTEXT000021519796",
-        "state": "VIGUEUR",
-        "date_of_signature": 1258585200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35826",
-        "title": " Arrêté du 19/12/08 fixant les règles générales et prescriptions techniques applicables aux installations classées soumises à autorisation sous la rubrique n° 1434-1 ",
-        "nor": "DEVP0827679A",
-        "classements": [{"regime": "A", "rubrique": "1434", "alinea": "1"}],
-        "cid": "JORFTEXT000019994396",
-        "state": "VIGUEUR",
-        "date_of_signature": 1229641200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "35844",
-        "title": " Arrêté du 12/10/11 relatif aux installations classées soumises à autorisation au titre de la rubrique 1434-2 de la législation des installations classées pour la protection de l’environnement ",
-        "nor": "DEVP1127666A",
-        "classements": [{"regime": "A", "rubrique": "1434", "alinea": "2"}],
-        "cid": "JORFTEXT000024778832",
-        "state": "VIGUEUR",
-        "date_of_signature": 1318370400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "37316",
-        "title": " Arrêté du 15/02/16 relatif aux installations de stockage de déchets de sédiments ",
-        "nor": "DEVP1519170A",
-        "classements": [
-            {"regime": "A", "rubrique": "2760", "alinea": "1"},
-            {"regime": "A", "rubrique": "2760", "alinea": "2"},
-        ],
-        "cid": "JORFTEXT000032282438",
-        "state": "VIGUEUR",
-        "date_of_signature": 1455490800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3497",
-        "title": " Arrêté du 26/03/12 relatif aux prescriptions générales applicables aux installations classées relevant du régime de l'enregistrement au titre de la rubrique n° 2710-2 (installations de collecte de déchets non dangereux apportés par leur producteur initial) de la nomenclature des installations classées pour la protection de l'environnement  ",
-        "nor": "DEVP1208907A",
-        "classements": [{"regime": "E", "rubrique": "2710", "alinea": "2"}],
-        "cid": "JORFTEXT000025634333",
-        "state": "VIGUEUR",
-        "date_of_signature": 1332712800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40759",
-        "title": " Arrêté du 06/06/18 relatif aux prescriptions générales applicables aux installations de transit, regroupement, tri ou préparation en vue de la réutilisation de déchets relevant du régime de l'enregistrement au titre de la rubrique n° 2711 (déchets d'équipements électriques et électroniques), 2713 (métaux ou déchets de métaux non dangereux, alliage de métaux ou déchets d'alliage de métaux non dangereux), 2714 (déchets non dangereux de papiers, cartons, plastiques, caoutchouc, textiles, bois) ou 2716 (déchets non dangereux non inertes) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1800801A",
-        "classements": [
-            {"regime": "E", "rubrique": "2711", "alinea": "1"},
-            {"regime": "E", "rubrique": "2713", "alinea": "1"},
-            {"regime": "E", "rubrique": "2714", "alinea": "1"},
-            {"regime": "E", "rubrique": "2716", "alinea": "1"},
-        ],
-        "cid": "JORFTEXT000037032383",
-        "state": "VIGUEUR",
-        "date_of_signature": 1528236000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "22371",
-        "title": " Arrêté du 26/11/12 relatif aux prescriptions générales applicables aux installations classées relevant du régime de l'enregistrement au titre de la rubrique n° 2712-1 (installation d'entreposage, dépollution, démontage ou découpage de véhicules terrestres hors d'usage) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1238447A",
-        "classements": [{"regime": "E", "rubrique": "2712", "alinea": "1"}],
-        "cid": "JORFTEXT000026695033",
-        "state": "VIGUEUR",
-        "date_of_signature": 1353884400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "40753",
-        "title": " Arrêté du 06/06/18 relatif aux prescriptions générales applicables aux installations d'entreposage, dépollution, démontage ou découpage de déchets issus de bateaux de plaisance ou de sport tels que définis à l'article R. 543-297 du code de l'environnement relevant du régime de l'enregistrement au titre de la rubrique n° 2712-3 de la nomenclature des installations classées pour la protection de l'environnement et modifiant l'arrêté du 26 novembre 2012 relatif aux prescriptions générales applicables aux installations classées relevant du régime de l'enregistrement au titre de la rubrique n° 2712-1 (installation d'entreposage, dépollution, démontage ou découpage de véhicules terrestres hors d'usage) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1800785A",
-        "classements": [{"regime": "E", "rubrique": "2712", "alinea": "3"}],
-        "cid": "JORFTEXT000037032238",
-        "state": "VIGUEUR",
-        "date_of_signature": 1528236000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "36244",
-        "title": " Arrêté du 02/10/15 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2731-1 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1427353A",
-        "classements": [{"regime": "E", "rubrique": "2731", "alinea": "1"}],
-        "cid": "JORFTEXT000031259868",
-        "state": "VIGUEUR",
-        "date_of_signature": 1443736800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41576",
-        "title": " Arrêté du 27/11/18 relatif aux prescriptions générales applicables aux installations de stockage de déchets non dangereux dans une implantation isolée telle que définie dans la directive 1999/31/CE relevant du régime de l'enregistrement au titre de la rubrique n° 2760-2a de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1800795A",
-        "classements": [{"regime": "E", "rubrique": "2760", "alinea": "2.a"}],
-        "cid": "JORFTEXT000037832596",
-        "state": "VIGUEUR",
-        "date_of_signature": 1543273200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4015",
-        "title": " Arrêté du 12/08/10 relatif aux prescriptions générales applicables aux installations classées de méthanisation relevant du régime de l'enregistrement au titre de la rubrique \" n° 2781\" de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1020761A",
-        "classements": [{"regime": "E", "rubrique": "2781", "alinea": "1"}],
-        "cid": "JORFTEXT000022727437",
-        "state": "VIGUEUR",
-        "date_of_signature": 1281564000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "27539",
-        "title": " Arrêté du 24/09/13 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2910-B de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "DEVP1300516A",
-        "classements": [{"regime": "E", "rubrique": "2910", "alinea": "B"}],
-        "cid": "JORFTEXT000028071164",
-        "state": "ABROGE",
-        "date_of_signature": 1379973600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3623",
-        "title": " Arrêté du 08/12/11 relatif aux prescriptions générales applicables aux installations relevant du régime de l’enregistrement au titre de la rubrique n° 2910-C de la nomenclature des installations classées pour la protection de l’environnement (installations de combustion consommant exclusivement du biogaz produit par une seule installation de méthanisation soumise à enregistrement sous la rubrique n° 2781-1) ",
-        "nor": "DEVP1132167A",
-        "classements": [{"regime": "E", "rubrique": "2910", "alinea": "C"}],
-        "cid": "JORFTEXT000025178828",
-        "state": "ABROGE",
-        "date_of_signature": 1323298800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41029",
-        "title": " Arrêté du 03/08/18 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de rubrique 2910 de la nomenclature des installations classées pour la protection de l'environnement (applicable à compter du 20 décembre 2018) ",
-        "nor": "TREP1726510A",
-        "classements": [
-            {"regime": "E", "rubrique": "2910", "alinea": "B"},
-            {"regime": "E", "rubrique": "2910", "alinea": "C"},
-        ],
-        "cid": "JORFTEXT000037284611",
-        "state": "VIGUEUR",
-        "date_of_signature": 1533247200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "43224",
-        "title": " Arrêté du 12/05/20 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2915 (Procédés de chauffage utilisant comme fluide caloporteur des corps organiques combustibles, lorsque la température d'utilisation est égale ou supérieure au point éclair des fluides) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1935396A",
-        "classements": [{"regime": "E", "rubrique": "2915", "alinea": "1"}],
-        "cid": "JORFTEXT000041875990",
-        "state": "VIGUEUR",
-        "date_of_signature": 1589234400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "43226",
-        "title": " Arrêté du 12/05/20 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2930 (Ateliers de réparation et d'entretien de véhicules et engins à moteur, y compris les activités de carrosserie et de tôlerie) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1935397A",
-        "classements": [
-            {"regime": "E", "rubrique": "2930", "alinea": "2"},
-            {"regime": "E", "rubrique": "2930", "alinea": "1"},
-        ],
-        "cid": "JORFTEXT000041876039",
-        "state": "VIGUEUR",
-        "date_of_signature": 1589234400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "43228",
-        "title": " Arrêté du 12/05/20 relatif aux prescriptions générales applicables aux installations relevant du régime de l'enregistrement au titre de la rubrique n° 2940 (Application, revêtement, laquage, stratification, imprégnation, cuisson, séchage de vernis, peinture, apprêt, colle, enduit, etc., sur support quelconque) de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1935398A",
-        "classements": [
-            {"regime": "E", "rubrique": "2940", "alinea": "1"},
-            {"regime": "E", "rubrique": "2940", "alinea": "2"},
-            {"regime": "E", "rubrique": "2940", "alinea": "3"},
-        ],
-        "cid": "JORFTEXT000041876133",
-        "state": "VIGUEUR",
-        "date_of_signature": 1589234400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "4989",
-        "title": " Arrêté du 12/12/05 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1310-2.c (Abrogé) ",
-        "nor": "DEVP0540470A",
-        "classements": [{"regime": "D", "rubrique": "1310", "alinea": "2.c", "state": "SUPPRIMEE"}],
-        "cid": "JORFTEXT000000263777",
-        "state": "ABROGE",
-        "date_of_signature": 1134342000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4003",
-        "title": " Arrêté du 30/08/10 relatif aux prescriptions applicables aux installations classées pour la protection de l’environnement soumises à déclaration sous la rubrique n° 1414-3 : Installations de remplissage ou de distribution de gaz inflammables liquéfiés : installations de remplissage de réservoirs alimentant des moteurs ou autres appareils d’utilisation comportant des organes de sécurité (jauges et soupapes) ",
-        "nor": "DEVP1020254A",
-        "classements": [{"regime": "D", "rubrique": "1414", "alinea": "3"}],
-        "cid": "JORFTEXT000022865723",
-        "state": "VIGUEUR",
-        "date_of_signature": 1283119200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "35823",
-        "title": " Arrêté du 19/12/08 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 1434 ",
-        "nor": "DEVP0827677A",
-        "classements": [{"regime": "D", "rubrique": "1434", "alinea": "1"}],
-        "cid": "JORFTEXT000020008930",
-        "state": "VIGUEUR",
-        "date_of_signature": 1229641200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4835",
-        "title": " Arrêté du 22/01/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2101-4 ",
-        "nor": "DEVP0700077A",
-        "classements": [{"regime": "D", "rubrique": "2101", "alinea": "4"}],
-        "cid": "JORFTEXT000000464327",
-        "state": "VIGUEUR",
-        "date_of_signature": 1169420400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5587",
-        "title": " Arrêté du 29/12/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2160 : \" Silos et installations de stockage de céréales, graines, produits alimentaires ou tout autre produit organique dégageant des poussières inflammables \" (Abrogé) ",
-        "nor": "ATEP9980030A",
-        "classements": [{"regime": "D", "rubrique": "2160", "alinea": "1"}],
-        "cid": "JORFTEXT000000208529",
-        "state": "ABROGE",
-        "date_of_signature": 914886000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4591",
-        "title": " Arrêté du 28/12/07 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2160 \" Silos et installations de stockage en vrac de céréales, grains, produits alimentaires ou tout produit organique dégageant des poussières inflammables, y compris les stockages sous tente ou structure gonflable \" ",
-        "nor": "DEVP0773639A",
-        "classements": [
-            {"regime": "D", "rubrique": "2160", "alinea": "1"},
-            {"regime": "D", "rubrique": "2160", "alinea": "2"},
-        ],
-        "cid": "JORFTEXT000018049814",
-        "state": "VIGUEUR",
-        "date_of_signature": 1198796400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "5455",
-        "title": " Arrêté du 18/12/00 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2160-2 : \" Silos et installations de stockage de céréales, grains, produits alimentaires ou tout autre produit organique dégageant des poussières inflammables \" (Abrogé) ",
-        "nor": "ATEP0100003A",
-        "classements": [{"regime": "D", "rubrique": "2160", "alinea": "2"}],
-        "cid": "JORFTEXT000000768178",
-        "state": "ABROGE",
-        "date_of_signature": 977094000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "42632",
-        "title": " Arrêté du 30/10/19 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration sous la rubrique n° 2210-3 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1931422A",
-        "classements": [{"regime": "D", "rubrique": "2210", "alinea": "3"}],
-        "cid": "JORFTEXT000039397758",
-        "state": "VIGUEUR",
-        "date_of_signature": 1572390000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5627",
-        "title": " Arrêté du 02/06/98 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2680-1 : Organismes génétiquement modifiés (Installations où sont mis en œuvre un processus de production industrielle ou commerciale des) ",
-        "nor": "ATEP9870205A",
-        "classements": [{"regime": "D", "rubrique": "2680", "alinea": "1"}],
-        "cid": "JORFTEXT000000194255",
-        "state": "VIGUEUR",
-        "date_of_signature": 896738400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3489",
-        "title": " Arrêté du 27/03/12 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2710-1 (Installations de collecte de déchets dangereux apportés par leur producteur initial)  ",
-        "nor": "DEVP1208904A",
-        "classements": [{"regime": "D", "rubrique": "2710", "alinea": "1"}],
-        "cid": "JORFTEXT000025687401",
-        "state": "VIGUEUR",
-        "date_of_signature": 1332799200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3491",
-        "title": " Arrêté du 27/03/12 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2710-2 (Installations de collecte de déchets non dangereux apportés par leur producteur initial)  ",
-        "nor": "DEVP1208913A",
-        "classements": [{"regime": "D", "rubrique": "2710", "alinea": "2"}],
-        "cid": "JORFTEXT000025687415",
-        "state": "VIGUEUR",
-        "date_of_signature": 1332799200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "3749",
-        "title": " Arrêté du 18/07/11 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2718 (installation de transit, regroupement ou tri de déchets dangereux ou de déchets contenant les substances dangereuses ou préparations dangereuses mentionnées à l’article R. 511-10 du code de l’environnement, à l’exclusion des installations visées aux rubriques 1313, 2710, 2711, 2712, 2717 et 2719) (applicable jusqu'au 30 juin 2018) ",
-        "nor": "DEVP1117061A",
-        "classements": [{"regime": "D", "rubrique": "2718"}],
-        "cid": "JORFTEXT000024433621",
-        "state": "ABROGE",
-        "date_of_signature": 1310940000,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41351",
-        "title": " Arrêté du 22/10/18 relatif aux prescriptions générales applicables aux installations relevant du régime de la déclaration avec contrôle périodique au titre de la rubrique n° 2731-3 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1813028A",
-        "classements": [{"regime": "D", "rubrique": "2731", "alinea": "3"}],
-        "cid": "JORFTEXT000037518915",
-        "state": "VIGUEUR",
-        "date_of_signature": 1540159200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "4221",
-        "title": " Arrêté du 10/11/09 relatif aux prescriptions générales applicables aux installations classées de méthanisation soumises à déclaration sous la rubrique n° 2781-1 ",
-        "nor": "DEVP0920876A",
-        "classements": [{"regime": "D", "rubrique": "2781", "alinea": "1"}],
-        "cid": "JORFTEXT000021334587",
-        "state": "VIGUEUR",
-        "date_of_signature": 1257807600,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "42121",
-        "title": " Arrêté du 8 mars 2019 relatif aux prescriptions générales applicables aux installations de transit, tri, regroupement de déchets contenant des PCB/PCT à une concentration supérieure à 50 ppm relevant du régime de la déclaration au titre de la rubrique n° 2792-1 de la nomenclature des installations classées pour la protection de l'environnement ",
-        "nor": "TREP1834476A",
-        "classements": [{"regime": "D", "rubrique": "2792", "alinea": "1"}],
-        "cid": "JORFTEXT000038528312",
-        "state": "VIGUEUR",
-        "date_of_signature": 1551999600,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "33745",
-        "title": " Arrêté du 16/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2793-1 ",
-        "nor": "DEVP1426102A",
-        "classements": [{"regime": "D", "rubrique": "2793", "alinea": "1"}],
-        "cid": "JORFTEXT000029964686",
-        "state": "VIGUEUR",
-        "date_of_signature": 1418684400,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "33747",
-        "title": " Arrêté du 16/12/14 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2793-2 ",
-        "nor": "DEVP1426105A",
-        "classements": [{"regime": "D", "rubrique": "2793", "alinea": "2"}],
-        "cid": "JORFTEXT000029964692",
-        "state": "VIGUEUR",
-        "date_of_signature": 1418684400,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "39938",
-        "title": " Arrêté du 21/11/17 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration sous la rubrique n° 2793-3a ",
-        "nor": "TREP1713996A",
-        "classements": [{"regime": "D", "rubrique": "2793", "alinea": "3"}],
-        "cid": "JORFTEXT000036126905",
-        "state": "VIGUEUR",
-        "date_of_signature": 1511218800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "3625",
-        "title": " Arrêté du 08/12/11 relatif aux prescriptions générales applicables aux installations classées soumises à déclaration sous la rubrique n° 2910-C de la nomenclature des installations classées pour la protection de l’environnement (installations de combustion consommant exclusivement du biogaz produit par une seule installation de méthanisation soumise à déclaration sous la rubrique n° 2781-1) ",
-        "nor": "DEVP1132166A",
-        "classements": [{"regime": "D", "rubrique": "2910", "alinea": "C"}],
-        "cid": "JORFTEXT000025014944",
-        "state": "ABROGE",
-        "date_of_signature": 1323298800,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "41027",
-        "title": " Arrêté du 03/08/18 relatif aux prescriptions générales applicables aux appareils de combustion, consommant du biogaz produit par des installations de méthanisation classées sous la rubrique n° 2781-1, inclus dans une installation de combustion classée pour la protection de l'environnement soumise à déclaration sous la rubrique n° 2910 (applicable à compter du 20 décembre 2018) ",
-        "nor": "TREP1726505A",
-        "classements": [{"regime": "D", "rubrique": "2910", "alinea": "A.2"}],
-        "cid": "JORFTEXT000037284579",
-        "state": "DELETED",
-        "reason_deleted": "Cet AM est assez particulier : il s'applique aux 2781 dans les 2910",
-        "date_of_signature": 1533247200,
-        "source": "AIDA",
-    },
-    {
-        "aida_page": "41025",
-        "title": " Arrêté du 03/08/18 relatif aux prescriptions générales applicables aux installations classées pour la protection de l'environnement soumises à déclaration au titre de la rubrique 2910 (applicable à compter du 20 décembre 2018) ",
-        "nor": "TREP1726498A",
-        "classements": [{"regime": "D", "rubrique": "2910", "alinea": "A.2"}],
-        "cid": "JORFTEXT000037284510",
-        "state": "VIGUEUR",
-        "date_of_signature": 1533247200,
-        "source": "LEGIFRANCE",
-    },
-    {
-        "aida_page": "5619",
-        "title": "Faux Arrêté du 01/02/21 relatif aux tests",
-        "nor": "FAKE_NOR",
-        "classements": [{"regime": "E", "rubrique": "5000", "alinea": "A.2"}],
-        "cid": "FAKE_CID",
-        "state": "VIGUEUR",
-        "date_of_signature": 1612195449,
-        "source": "LEGIFRANCE",
-    },
-]
